refactor(daemon): extract the cross-machine pool transport out of lib.rs
#3079 landed ~2.7k lines of cross-machine memory-pool transport inline in
`daemon/src/lib.rs`, taking it to 14.2k lines. Moves it verbatim into
`memory_pool/`, split by plane:

  auth.rs        shared-token handshake for the direct-TCP plane
  mirror.rs      DORADMA header, seqlock, /dev/shm mirror segments
  data_plane.rs  direct-TCP listener and sender
  control.rs     zenoh control events, pending-ack registries

No behavior change: code moves, visibility widens to `pub(crate)`, the
three pool test modules move with the code they cover. lib.rs drops from
14,202 to 11,384 lines and back to **zero `unsafe`** — all 47 occurrences
belong to the pool transport. 100 test functions before and after; 251
daemon tests pass.
