+++ binaries/daemon/src/lib.rs
@@ -413,1456 +413,8 @@ pub(crate) struct PendingDestroy {
     wait: shutdown::DestroyWait,
 }
 
-// ---------------------------------------------------------------------------
-// Cross-machine memory pool data plane (ports the pre-tensor-pool branch;
-// see `docs/design.md` in the hetero-pool PR). All of it is opt-in behind
-// `DORA_MEMORY_POOL_CROSS_MACHINE=1`; a daemon without the env never opens
-// the direct-TCP data port and never mirrors remote pools.
-// ---------------------------------------------------------------------------
-
-/// Opt-in switch for the whole cross-machine memory-pool data plane.
-/// Off by default: no mirror segments, no direct-TCP listener, no
-/// RegisterPool/RegisterCrossMachinePool processing. Deployments that
-/// enable it must set the same value on every participating daemon.
-const CROSS_MACHINE_ENV: &str = "DORA_MEMORY_POOL_CROSS_MACHINE";
-
-fn cross_machine_enabled() -> bool {
-    matches!(
-        std::env::var(CROSS_MACHINE_ENV).as_deref(),
-        Ok("1") | Ok("true") | Ok("TRUE") | Ok("yes")
-    )
-}
-
-/// Auth token for the direct-TCP data plane. When set (on every
-/// participating daemon), each direct-TCP connection performs a token
-/// handshake before the first data frame; a mismatched or missing token
-/// drops the connection (the origin degrades to the zenoh relay). Unset
-/// = no handshake (the daemon then relies on the port's reachability).
-/// The token is NOT injected into spawned nodes (the handshake runs
-/// entirely in the daemon; node-side code never reads it — injecting it
-/// would leak the shared secret into every user node process, bot
-/// review 5301862843).
-///
-/// Scope: the token authenticates the direct-TCP data plane only. The
-/// zenoh control/relay plane (RegisterPool/FreePool/MemoryPoolWrite over
-/// `dataflow_memory_pool_topic`) is unauthenticated — dora's mesh-daemon
-/// model treats peers as trusted. When this token is relied on over an
-/// untrusted link (edge↔cloud WAN), the zenoh transport itself must be
-/// secured (e.g. TLS via the zenoh config), or the control plane must be
-/// firewalled to trusted peers.
-const CROSS_DATA_AUTH_TOKEN_ENV: &str = "DORA_MEMORY_POOL_AUTH_TOKEN";
-
-fn cross_data_auth_token() -> Option<String> {
-    std::env::var(CROSS_DATA_AUTH_TOKEN_ENV)
-        .ok()
-        .filter(|t| !t.is_empty())
-}
-
-/// Handshake frame: `[8-byte magic][u32 token_len][token]`, answered with
-/// a single byte (1 = accepted, 0 = rejected + connection close).
-const CROSS_DATA_AUTH_MAGIC: [u8; 9] = *b"DORA_AUTH";
-const AUTH_OK: u8 = 1;
-const AUTH_FAIL: u8 = 0;
-
-/// Send-side handshake: runs once right after connecting, before any
-/// data frame. The peer closes the connection on rejection, which
-/// surfaces here as a read error or `AUTH_FAIL` byte.
-async fn auth_handshake_send(
-    stream: &mut tokio::net::TcpStream,
-    token: &str,
-) -> Result<(), String> {
-    use tokio::io::AsyncWriteExt;
-    stream
-        .write_all(&CROSS_DATA_AUTH_MAGIC)
-        .await
-        .map_err(|e| format!("auth handshake write failed: {e}"))?;
-    let token = token.as_bytes();
-    stream
-        .write_all(&(token.len() as u32).to_be_bytes())
-        .await
-        .map_err(|e| format!("auth handshake write failed: {e}"))?;
-    stream
-        .write_all(token)
-        .await
-        .map_err(|e| format!("auth handshake write failed: {e}"))?;
-    stream
-        .flush()
-        .await
-        .map_err(|e| format!("auth handshake flush failed: {e}"))?;
-    let mut reply = [0u8; 1];
-    tokio::io::AsyncReadExt::read_exact(stream, &mut reply)
-        .await
-        .map_err(|e| format!("auth handshake reply read failed: {e}"))?;
-    if reply[0] != AUTH_OK {
-        return Err("peer rejected the auth token".to_string());
-    }
-    Ok(())
-}
-
-/// Receive-side handshake: validates the peer's token. Bounded by
-/// [`CROSS_DATA_READ_TIMEOUT`] so a connection that never sends the
-/// handshake cannot pin a task (or an fd) forever.
-async fn auth_handshake_verify(stream: &mut tokio::net::TcpStream) -> Result<(), String> {
-    use tokio::io::{AsyncReadExt, AsyncWriteExt};
-    let mut magic = [0u8; 9];
-    stream
-        .read_exact(&mut magic)
-        .await
-        .map_err(|e| format!("auth handshake read failed: {e}"))?;
-    if magic != CROSS_DATA_AUTH_MAGIC {
-        let _ = stream.write_all(&[AUTH_FAIL]).await;
-        return Err("bad auth handshake magic (peer without the shared token?)".to_string());
-    }
-    let mut len_bytes = [0u8; 4];
-    stream
-        .read_exact(&mut len_bytes)
-        .await
-        .map_err(|e| format!("auth handshake read failed: {e}"))?;
-    let len = u32::from_be_bytes(len_bytes) as usize;
-    if len > 4096 {
-        let _ = stream.write_all(&[AUTH_FAIL]).await;
-        return Err(format!("auth token too long ({len} bytes)"));
-    }
-    let mut token = vec![0u8; len];
-    stream
-        .read_exact(&mut token)
-        .await
-        .map_err(|e| format!("auth handshake read failed: {e}"))?;
-    let expected = cross_data_auth_token().unwrap_or_default();
-    // Constant-time comparison (parity with the message-layer auth): the
-    // handshake token is a shared deployment secret, so a timing side
-    // channel on the comparison would leak it byte by byte.
-    let token_ok = token.len() == expected.len()
-        && token
-            .iter()
-            .zip(expected.as_bytes())
-            .fold(0u8, |acc, (a, b)| acc | (a ^ b))
-            == 0;
-    if token_ok {
-        stream
-            .write_all(&[AUTH_OK])
-            .await
-            .map_err(|e| format!("auth handshake reply write failed: {e}"))?;
-        Ok(())
-    } else {
-        let _ = stream.write_all(&[AUTH_FAIL]).await;
-        Err("peer auth token mismatch".to_string())
-    }
-}
-
-/// Gate a fresh direct-TCP connection on the auth handshake.
-///
-/// Returns `Ok(())` when the peer passed the handshake (or no token is
-/// configured — a token-less daemon accepts token-less peers), and
-/// `Err(reason)` when the connection must be dropped *before* any frame
-/// is served. This is the listener's consumption of
-/// [`auth_handshake_verify`]; factored out so the accept loop's
-/// reject-on-mismatch behavior is testable through a real socket.
-///
-/// Note the nested-result flatten: `timeout(...).await` is
-/// `Result<Result<(), String>, Elapsed>` — an `Err` from the verify
-/// (bad magic, over-long or wrong token) is the *inner* result and must
-/// be rejected here, not just the outer timeout.
-async fn cross_data_auth_gate(stream: &mut tokio::net::TcpStream) -> Result<(), String> {
-    if cross_data_auth_token().is_none() {
-        return Ok(());
-    }
-    match tokio::time::timeout(CROSS_DATA_READ_TIMEOUT, auth_handshake_verify(stream)).await {
-        Ok(Ok(())) => Ok(()),
-        Ok(Err(e)) => Err(format!("auth handshake rejected: {e}")),
-        Err(_) => Err("auth handshake timed out".to_string()),
-    }
-}
-
-/// Per-pool synchronous write lock for the zenoh-relay data plane.
-/// Keyed by `(dataflow id, pool id)` like the direct-TCP twin
-/// `CROSS_POOL_WRITE_LOCKS_ASYNC`: pool ids repeat across dataflows (each
-/// node process restarts its counter), and a bare pool-id key would
-/// serialize writes to different segments of concurrent dataflows.
-type CrossPoolWriteLocksSync = std::sync::Mutex<
-    std::collections::HashMap<(Uuid, String), std::sync::Arc<std::sync::Mutex<()>>>,
->;
-static CROSS_POOL_WRITE_LOCKS: std::sync::LazyLock<CrossPoolWriteLocksSync> =
-    std::sync::LazyLock::new(CrossPoolWriteLocksSync::default);
-
-/// Async per-pool write lock for the direct-TCP data plane: a
-/// `std::sync::MutexGuard` cannot be held across an `.await`, so the
-/// receive-into-mirror path (which reads the stream while holding the
-/// per-pool serialization lock) uses an async mutex instead. Serializes
-/// concurrent direct writes to the **same pool of the same dataflow**
-/// across connections. Keyed by `(dataflow id, pool id)` like the other
-/// cross-write state: pool ids repeat across dataflows (each node process
-/// restarts its counter), and a bare pool-id key would needlessly
-/// serialize writes to *different* segments of concurrent dataflows.
-/// Per-plane only: the relay path serializes on the separate sync
-/// `CROSS_POOL_WRITE_LOCKS`; cross-plane exclusion is provided by the
-/// node's turn-based write cadence, not by these locks (bot review
-/// 5306582566; the dual-lock split is a decided design).
-type CrossPoolWriteLocks = tokio::sync::Mutex<
-    std::collections::HashMap<(Uuid, String), std::sync::Arc<tokio::sync::Mutex<()>>>,
->;
-static CROSS_POOL_WRITE_LOCKS_ASYNC: std::sync::LazyLock<CrossPoolWriteLocks> =
-    std::sync::LazyLock::new(CrossPoolWriteLocks::default);
-
-// DORADMA shmem layout — must match the node API exactly
-// (apis/python/node/src/lib.rs): [magic:8][json_len:8][data_offset:8]
-// [ipc_present:8][ipc_handle:64][write_gen:8 @96][reserved:152][json:256]
-// [data:data_offset]. write_gen is the seqlock generation: even =
-// complete, odd = write in progress.
-const DORADMA_HEADER_SIZE: usize = 256;
-const DORADMA_MAGIC: &[u8; 8] = b"DORADMA\x00";
-
-/// Read 8 consecutive bytes from `ptr` as a little-endian u64. Same
-/// implementation as the node API's `read_header_u64` — the wire layout
-/// of the DORADMA header must be identical on both sides.
-fn read_header_u64(ptr: *const u8) -> u64 {
-    const { assert!(std::mem::size_of::<u64>() == 8) };
-    let mut buf = [0u8; 8];
-    unsafe { std::ptr::copy_nonoverlapping(ptr, buf.as_mut_ptr(), 8) };
-    u64::from_le_bytes(buf)
-}
-
-/// Write 8 bytes as a little-endian u64 at `ptr`. Mirror of the node
-/// API's header writes (`json_len_le` / `data_off_le` byte copies).
-fn write_header_u64(ptr: *mut u8, value: u64) {
-    const { assert!(std::mem::size_of::<u64>() == 8) };
-    let le = value.to_le_bytes();
-    unsafe { std::ptr::copy_nonoverlapping(le.as_ptr(), ptr, 8) };
-}
-
-/// Begins a memory-pool seqlock write at `gen_ptr` (header offset 96)
-/// **if the generation is even**: marks the generation odd (write in
-/// progress) and returns the even pre-write generation. If the
-/// generation is already odd (leftover from a previous failed write),
-/// the increment is skipped and the previous even generation is
-/// returned, so `seqlock_end`'s `pre + 2` always produces an even
-/// generation. Bit-identical to the node API's `seqlock_begin_if_even`.
-unsafe fn seqlock_begin_if_even(gen_ptr: *mut u64) -> u64 {
-    unsafe {
-        let cur = std::ptr::read_volatile(gen_ptr);
-        if cur.is_multiple_of(2) {
-            std::ptr::write_volatile(gen_ptr, cur + 1);
-            std::sync::atomic::fence(std::sync::atomic::Ordering::Release);
-        }
-        cur & !1 // always return the even baseline
-    }
-}
-
-/// Closes a memory-pool seqlock write (header offset 96): publishes
-/// `pre_write_gen + 2` (even = complete) when the copy succeeded, or
-/// rolls back to `pre_write_gen` when it failed. Bit-identical to the
-/// node API's `seqlock_end`.
-unsafe fn seqlock_end(gen_ptr: *mut u64, pre_write_gen: u64, copy_ok: bool) {
-    unsafe {
-        if copy_ok {
-            std::ptr::write_volatile(gen_ptr, pre_write_gen.wrapping_add(2));
-        } else {
-            std::ptr::write_volatile(gen_ptr, pre_write_gen);
-        }
-        std::sync::atomic::fence(std::sync::atomic::Ordering::Release);
-    }
-}
-
-/// Remove stale segments left on this machine by dataflows whose daemon
-/// was killed without running shutdown cleanup. Only segments carrying
-/// THIS machine's id prefix (`dora_pool_{machine_id}_...`) are touched: a
-/// daemon restart implies its own dataflows died (nodes are daemon
-/// children), and sibling daemons on the same host use their own prefixes
-/// — so nothing live is ever unlinked. Machine-qualified LOCAL pool
-/// segments (the python side now qualifies auto names with
-/// `DORA_MACHINE_ID`) are swept here too — they are attributable to this
-/// machine and can only be leftovers of this daemon's own dead dataflows.
-#[cfg(target_os = "linux")]
-fn cleanup_orphan_mirrors(machine_id: &str) -> usize {
-    let prefix = format!("dora_pool_{machine_id}_");
-    let Ok(entries) = std::fs::read_dir("/dev/shm") else {
-        return 0;
-    };
-    let mut removed = 0;
-    for entry in entries.flatten() {
-        let name = entry.file_name();
-        let name = name.to_string_lossy();
-        if name.starts_with(&prefix) {
-            match std::fs::remove_file(entry.path()) {
-                Ok(()) => {
-                    tracing::info!("memory pool: removed orphan mirror segment {name}");
-                    removed += 1;
-                }
-                Err(e) => tracing::warn!("memory pool: failed to remove orphan mirror {name}: {e}"),
-            }
-        }
-    }
-    if removed > 0 {
-        tracing::info!("memory pool: cleaned {removed} orphan mirror segment(s)");
-    }
-    removed
-}
-
-/// The pool descriptor replicated into this daemon's extension table when
-/// a cross-machine pool is mirrored (RegisterPool). The sender's node
-/// stores the descriptor on its own daemon only; without this replication
-/// a receiver's daemon query (`load_metadata`) misses, so a GPU receiver
-/// never obtains the daemon-trusted size it requires before HtoD staging
-/// (`GPU_BUF_SIZES`) and every read fails as "pool does not exist". CPU
-/// receivers never notice — their fast path reads the mirror header
-/// directly — which is why CPU-only deployments cannot trip this.
-///
-/// Field set mirrors the python transport's `seam::store` descriptor so
-/// the receiver-side consumers (`load_metadata`, `GPU_BUF_SIZES`) find
-/// the same keys. `pinned_type` carries the receiver device ("cuda:0")
-/// exactly like the mirror header JSON (same input, same value).
-/// NOT cfg-gated: it only builds a JSON descriptor (no /dev/shm access),
-/// and the RegisterPool handler calls it unconditionally — gating it
-/// broke the macOS daemon build (何勇 review 5302853212).
-fn build_mirror_descriptor(
-    size: usize,
-    dtype: &str,
-    shape: &[i64],
-    mirror_shmem_name: &str,
-    device: &str,
-) -> Vec<u8> {
-    let mut params = MetadataParameters::new();
-    params.insert(
-        "size".to_string(),
-        dora_message::metadata::Parameter::Integer(size as i64),
-    );
-    params.insert(
-        "dtype".to_string(),
-        dora_message::metadata::Parameter::String(dtype.to_string()),
-    );
-    params.insert(
-        "shape".to_string(),
-        dora_message::metadata::Parameter::ListInt(shape.to_vec()),
-    );
-    params.insert(
-        "shared_memory_name".to_string(),
-        dora_message::metadata::Parameter::String(mirror_shmem_name.to_string()),
-    );
-    params.insert(
-        "pinned_type".to_string(),
-        dora_message::metadata::Parameter::String(device.to_string()),
-    );
-    params.insert(
-        "ipc_present".to_string(),
-        dora_message::metadata::Parameter::Bool(false),
-    );
-    serde_json::to_vec(&params).unwrap_or_default()
-}
-
-/// Create a CPU DORADMA pool mirror on this machine. Mirrors the node
-/// API's register_memory_pool shmem layout. The generation is
-/// initialized to an odd (in-progress) value with all-zero data so
-/// receivers do not read the empty segment as a valid frame before the
-/// first direct write lands (the reader retries while the generation is
-/// odd). The segment is deliberately left owner-less (`set_owner(false)`)
-/// so the /dev/shm name survives the handle drop — on Linux a created
-/// (owner) Shmem shm_unlinks on drop, which would remove the name local
-/// receivers open for the zero-copy fast path.
-#[allow(clippy::too_many_arguments)]
-fn create_cross_pool_shmem(
-    dataflow_id: &Uuid,
-    machine_id: &str,
-    shared_memory_id: &str,
-    size: usize,
-    dtype: &str,
-    shape: &[i64],
-    device: &str,
-    sender_shmem: Option<&str>,
-) -> eyre::Result<()> {
-    // Machine-qualified OS id: the mirror lives on the target machine's
-    // /dev/shm, which on a dual-daemon test host is the SAME namespace as
-    // the sender's local pool. An unqualified id would collide with the
-    // sender's local segment (create fails with EEXIST) whenever both
-    // daemons run on one host.
-    let shmem_name = TensorPoolManager::cross_pool_shmem_name(
-        machine_id,
-        &dataflow_id.to_string(),
-        shared_memory_id,
-    )
-    .ok_or_else(|| eyre::eyre!("invalid pool id: {shared_memory_id}"))?;
-    // `device` is the receiver's device (the mirror's consumer): the
-    // sender relays it in RegisterPool. A GPU receiver ("cuda:0") reads
-    // the mirror's CPU data region and stages it HtoD into its own GPU
-    // buffer; "cpu" readers consume the data region directly.
-    // `dtype`/`device` arrive from the remote RegisterPool event
-    // (untrusted cross-machine strings) — build the header JSON with
-    // serde_json so quotes/backslashes are escaped instead of corrupting
-    // or injecting into the parsed structure.
-    // `sender_shmem`: the sender's local segment name, relayed so a
-    // same-host reader can open it directly (the direct-TCP zero-copy
-    // fast path — the mirror is only a fallback that would otherwise
-    // serve stale frames, since the origin skips the per-frame push for
-    // same-host pools). Absent on cross-host mirrors, where readers
-    // consume the mirror's own data region.
-    let mut header_json = serde_json::json!({
-        "size": size,
-        "dtype": dtype,
-        "shape": shape,
-        "pinned_type": device,
-    });
-    if let Some(name) = sender_shmem {
-        header_json["sender_shmem"] = serde_json::Value::String(name.to_string());
-    }
-    let json = serde_json::to_string(&header_json)
-        .map_err(|e| eyre::eyre!("failed to serialize mirror header JSON: {e}"))?;
-    let data_offset = DORADMA_HEADER_SIZE + json.len();
-    let make_conf = || ShmemConf::new().os_id(&shmem_name).size(size + data_offset);
-    let mut shmem = match make_conf().create() {
-        Ok(s) => s,
-        Err(e) => {
-            // EEXIST: either a stale mirror left by a daemon that was
-            // killed without running shutdown cleanup, or — on a retried
-            // registration whose first ack was lost — this daemon's own
-            // live mirror created moments ago. Unlinking the latter
-            // strands receivers that already mapped the old inode (they
-            // read a frozen frame forever while writes land in the new
-            // one). Distinguish by segment size: a live mirror from this
-            // same (dataflow, pool) registration has exactly this size
-            // and its header is already written — treat it as idempotent
-            // success; anything else is stale and safe to replace (bot
-            // review 5307022693).
-            let shm_path = format!("/dev/shm/{shmem_name}");
-            if let Ok(meta) = std::fs::metadata(&shm_path)
-                && meta.len() == (size + data_offset) as u64
-            {
-                tracing::info!(
-                    "memory pool: mirror {shmem_name} already exists at the registered size \
-                     — idempotent success (registration retried after a lost ack)"
-                );
-                return Ok(());
-            }
-            if std::path::Path::new(&shm_path).exists() {
-                tracing::warn!("memory pool: stale mirror {shmem_name} exists, replacing");
-                std::fs::remove_file(&shm_path)
-                    .map_err(|e| eyre::eyre!("remove stale mirror {shmem_name}: {e}"))?;
-                make_conf()
-                    .create()
-                    .map_err(|re| eyre::eyre!("recreate mirror {shmem_name} after unlink: {re}"))?
-            } else {
-                return Err(eyre::eyre!("create shmem: {e}"));
-            }
-        }
-    };
-    unsafe {
-        let ptr = shmem.as_ptr();
-        std::ptr::copy_nonoverlapping(DORADMA_MAGIC.as_ptr(), ptr, 8);
-        write_header_u64(ptr.add(8), json.len() as u64);
-        write_header_u64(ptr.add(16), data_offset as u64);
-        std::ptr::copy_nonoverlapping(json.as_ptr(), ptr.add(DORADMA_HEADER_SIZE), json.len());
-        // Odd generation = write in progress. The mirror starts with
-        // all-zero data; an even (complete) generation would let a
-        // receiver read that as a valid frame before the first direct
-        // write, so begin odd and let the first `seqlock_end` publish
-        // the first even (complete) generation.
-        write_header_u64(ptr.add(96), 1);
-    }
-    shmem.set_owner(false);
-    Ok(())
-}
-
-/// Write tensor bytes into a mirrored cross-machine pool under the
-/// DORADMA seqlock protocol (odd gen during write, even after).
-///
-/// A 61.44MB mirror write is a 10-30ms synchronous memcpy, so callers
-/// must not run it on the event loop — spawn it (see the MemoryPoolWrite
-/// handler). Not async: there is nothing to await, the work is the copy.
-/// Returns whether the mirror write completed (the caller publishes the
-/// commit ack with this outcome).
-fn write_cross_pool_data(
-    dataflow_id: &Uuid,
-    machine_id: &str,
-    shared_memory_id: &str,
-    tensor_data: &[u8],
-    size: usize,
-) -> bool {
-    // Serialise concurrent relay-path writes to the same pool: two
-    // overlapping memcpys would interleave bytes and leave a mixed frame
-    // that the seqlock (odd = in-progress) cannot detect once both
-    // writers have completed an even generation. Per-pool lock, held
-    // across the whole write (open + seqlock begin + copy + end).
-    // NOTE: this lock is per-plane — the direct-TCP path serializes on
-    // the separate `CROSS_POOL_WRITE_LOCKS_ASYNC`. Cross-plane exclusion
-    // relies on the node's turn-based write cadence (each write awaits
-    // its seq-matched ack before the next), which the design assumes
-    // (bot review 5306582566; the dual-lock split is a decided design).
-    let write_lock = {
-        let mut locks = CROSS_POOL_WRITE_LOCKS
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-        let key = (*dataflow_id, shared_memory_id.to_string());
-        match locks.get(&key) {
-            Some(lock) => lock.clone(),
-            None => {
-                let lock = std::sync::Arc::new(std::sync::Mutex::new(()));
-                locks.insert(key, lock.clone());
-                lock
-            }
-        }
-    };
-    let _guard = write_lock.lock().unwrap_or_else(|e| e.into_inner());
-    // Must match the machine-qualified id used by `create_cross_pool_shmem`.
-    let Some(shmem_name) = TensorPoolManager::cross_pool_shmem_name(
-        machine_id,
-        &dataflow_id.to_string(),
-        shared_memory_id,
-    ) else {
-        tracing::warn!("memory pool: invalid pool id {shared_memory_id}, dropping frame");
-        return false;
-    };
-    let Ok(shmem) = ShmemConf::new().os_id(&shmem_name).open() else {
-        tracing::warn!(
-            "memory pool: pool {shared_memory_id} missing at write \
-             (sync register should have prevented this), dropping frame"
-        );
-        return false;
-    };
-    let shmem_ptr = shmem.as_ptr();
-    // Guard against a corrupt/truncated header before any pointer math.
-    let magic = unsafe { std::slice::from_raw_parts(shmem_ptr, 8) };
-    if magic != DORADMA_MAGIC {
-        tracing::warn!("memory pool: {shared_memory_id} header magic mismatch, dropping frame");
-        return false;
-    }
-    // The payload must cover the registered pool size recorded in the
-    // mirror header — a short payload would otherwise publish an even
-    // generation over a truncated tensor whose tail `[copy_len..]` is
-    // stale, and readers accept it as complete (the same gap the direct
-    // path's registered-size check closes; the zenoh relay is the active
-    // data plane whenever the direct endpoint is unavailable, not just a
-    // fallback).
-    let registered_size = {
-        let json_len = unsafe { read_header_u64(shmem_ptr.add(8)) } as usize;
-        let json_bytes =
-            unsafe { std::slice::from_raw_parts(shmem_ptr.add(DORADMA_HEADER_SIZE), json_len) };
-        serde_json::from_slice::<serde_json::Value>(json_bytes)
-            .ok()
-            .and_then(|v| v.get("size").and_then(|s| s.as_u64()))
-            .unwrap_or(0) as usize
-    };
-    if tensor_data.len() != registered_size || size != registered_size {
-        tracing::warn!(
-            "memory pool: {shared_memory_id} payload {} bytes / declared {size} \
-             != registered pool size {registered_size}, dropping frame",
-            tensor_data.len()
-        );
-        return false;
-    }
-    let data_offset = unsafe { read_header_u64(shmem_ptr.add(16)) } as usize;
-    let copy_len = tensor_data.len().min(size);
-    // Checked add: a corrupt header's data_offset could otherwise wrap
-    // `data_offset + copy_len` past the bounds check.
-    let Some(end) = data_offset.checked_add(copy_len) else {
-        tracing::warn!(
-            "memory pool: {shared_memory_id} data_offset {data_offset} + {copy_len} overflows usize, dropping frame"
-        );
-        return false;
-    };
-    if end > shmem.len() {
-        tracing::warn!(
-            "memory pool: {shared_memory_id} data_offset {data_offset} + {copy_len} exceeds shmem size {}, dropping frame",
-            shmem.len()
-        );
-        return false;
-    }
-    unsafe {
-        let gen_ptr = shmem_ptr.add(96) as *mut u64;
-        let pre = seqlock_begin_if_even(gen_ptr);
-        std::ptr::copy_nonoverlapping(tensor_data.as_ptr(), shmem_ptr.add(data_offset), copy_len);
-        seqlock_end(gen_ptr, pre, true);
-    }
-    true
-}
-
-/// Direct-TCP cross-machine data plane.
-///
-/// Frame: `[u32 magic][16-byte dataflow UUID][u32 pool_id_len][pool_id]
-/// [u64 seq][u64 size][size bytes of tensor data]`. The mirror daemon
-/// reads the tensor straight into the mirror segment's data region
-/// (zero user-space copies on the receive side); the origin pays a single
-/// user-space copy (segment → send buffer). The commit ack travels over
-/// zenoh (existing `MemoryPoolWriteAck` machinery) so the origin's
-/// pending resolution is unchanged. When `DORA_MEMORY_POOL_AUTH_TOKEN` is
-/// set, a token handshake precedes the first frame on every connection
-/// (see [`auth_handshake_send`]/[`auth_handshake_verify`]).
-const CROSS_DATA_MAGIC: u32 = 0xD0A0_0011;
-
-/// Keeps a mirror mapping alive while the direct-TCP payload is read
-/// straight into its data region. Only Send-safe values (plain addresses
-/// and the keep-alive `Shmem`) cross the `.await` in
-/// [`serve_cross_data_frame`]; the mapping itself is process-wide and
-/// thread-agnostic (same pattern as the python extension's `PoolSlot`).
-struct DirectMirrorWriter {
-    _shmem: shared_memory_extended::Shmem,
-    data_addr: usize,
-    data_len: usize,
-    gen_addr: usize,
-    pre: u64,
-}
-
-// SAFETY: the wrapper owns the mapping (`_shmem` keeps it alive until
-// `finish`/drop) and carries only plain addresses across awaits. Moving
-// the wrapper between threads never invalidates the mapping (mmap is
-// process-wide); no thread-bound state is involved.
-unsafe impl Send for DirectMirrorWriter {}
-
-impl DirectMirrorWriter {
-    /// Begin a seqlock write: computes the data region address from the
-    /// (already validated) header and takes the odd (in-progress)
-    /// generation.
-    fn new(shmem: shared_memory_extended::Shmem, data_offset: usize, size: usize) -> Self {
-        let shmem_ptr = shmem.as_ptr();
-        let gen_addr = unsafe { shmem_ptr.add(96) } as usize;
-        let pre = unsafe { seqlock_begin_if_even(gen_addr as *mut u64) };
-        Self {
-            _shmem: shmem,
-            data_addr: unsafe { shmem_ptr.add(data_offset) } as usize,
-            data_len: size,
-            gen_addr,
-            pre,
-        }
-    }
-
-    /// The mirror's data region, exactly `size` bytes (validated against
-    /// the segment length by the caller before construction).
-    fn data_slice_mut(&mut self) -> &mut [u8] {
-        unsafe { std::slice::from_raw_parts_mut(self.data_addr as *mut u8, self.data_len) }
-    }
-
-    /// Complete the seqlock write (even generation).
-    fn finish(self) {
-        unsafe { seqlock_end(self.gen_addr as *mut u64, self.pre, true) };
-    }
-}
-
-// NOTE on the failed-write path (no `Drop` rollback here): a payload
-// read that fails mid-frame leaves the generation odd (in-progress).
-// That is deliberate — readers reject the torn frame (they never see
-// half-written data), and the next full write self-heals the segment
-// (`seqlock_begin_if_even` finds the odd generation, keeps it, writes
-// the full frame, and publishes the even one). Rolling the generation
-// back to `pre` instead would *mark the torn bytes as a complete frame*,
-// which is worse than blocking. The origin fails fast through
-// `MemoryPoolWriteAck { ok: false }` (see `serve_cross_data_frame`).
-
-/// Port for the mirror daemon's direct-TCP data listener. Overridable for
-/// deployment (e.g. the rendezvous machine must publish this port to the
-/// origin machine on a routed WAN link).
-const CROSS_DATA_PORT_ENV: &str = "DORA_MEMORY_POOL_DATA_PORT";
-const CROSS_DATA_PORT_DEFAULT: u16 = 7410;
-/// Bind address for the direct-TCP data listener. Overridable for
-/// deployments where the listener must not (or cannot) sit on every
-/// interface — a host with a site firewall that only opens a specific
-/// interface, or a shared machine where port 7410 on `0.0.0.0` would
-/// collide with a sibling daemon.
-const CROSS_DATA_BIND_ENV: &str = "DORA_MEMORY_POOL_DATA_BIND";
-const CROSS_DATA_BIND_DEFAULT: &str = "0.0.0.0";
-
-/// Read and serve one direct-TCP data frame: write the payload straight
-/// into the mirror segment's data region (under the per-pool lock and
-/// seqlock) and publish the zenoh commit ack. Returns `Ok(true)` for the
-/// next frame, `Ok(false)` on clean EOF.
-async fn serve_cross_data_frame(
-    tensor_pool: &TensorPoolManager,
-    machine_id: &str,
-    session: &zenoh::Session,
-    clock: &Arc<HLC>,
-    shm_provider: Option<&ShmProvider<PosixShmProviderBackend>>,
-    stream: &mut tokio::net::TcpStream,
-) -> Result<bool, String> {
-    // Bound the whole frame read: a peer that sends a header with a large
-    // `size` and then stalls would otherwise hold the per-pool lock and
-    // leave the seqlock odd indefinitely, wedging every subsequent write
-    // to that pool. On timeout the connection is dropped; the lock guard
-    // drops with the cancelled future, and the odd generation marks the
-    // frame torn (next write self-heals).
-    let frame = match tokio::time::timeout(
-        CROSS_DATA_READ_TIMEOUT,
-        handle_cross_data_frame(stream, tensor_pool, machine_id),
-    )
-    .await
-    {
-        Ok(frame) => frame,
-        Err(_) => {
-            return Err(
-                "memory pool: direct-TCP data connection stalled (frame read timeout)".to_string(),
-            );
-        }
-    };
-    match frame {
-        Ok(Some((dataflow_id, shared_memory_id, seq))) => {
-            // Remote commit ack via zenoh (the origin's pending reply
-            // waits on it).
-            publish_memory_pool_event(
-                session,
-                clock,
-                &dataflow_id,
-                &InterDaemonEvent::MemoryPoolWriteAck {
-                    dataflow_id,
-                    shared_memory_id,
-                    seq,
-                    ok: true,
-                    error: None,
-                },
-                shm_provider,
-            )
-            .await
-            .map_err(|e| format!("failed to publish MemoryPoolWriteAck: {e}"))?;
-            Ok(true)
-        }
-        Ok(None) => Ok(false),
-        Err(err) => {
-            // The frame's identity is known: fail the origin's pending
-            // write fast (mirror could not write the frame — read error,
-            // segment missing, bounds violation) instead of making it
-            // wait out the commit-ack timeout.
-            if let Some((dataflow_id, shared_memory_id, seq)) = err.ack
-                && let Err(e) = publish_memory_pool_event(
-                    session,
-                    clock,
-                    &dataflow_id,
-                    &InterDaemonEvent::MemoryPoolWriteAck {
-                        dataflow_id,
-                        shared_memory_id,
-                        seq,
-                        ok: false,
-                        error: Some(err.message.clone()),
-                    },
-                    shm_provider,
-                )
-                .await
-            {
-                tracing::warn!(
-                    "memory pool: failed to publish failed-write MemoryPoolWriteAck: {e}"
-                );
-            }
-            Err(err.message)
-        }
-    }
-}
-
-/// Frame-level error from [`handle_cross_data_frame`]. Carries the
-/// frame's identity `(dataflow id, pool id, seq)` once the header has
-/// been parsed, so the caller can publish `MemoryPoolWriteAck { ok: false }`
-/// and let the origin fail fast instead of waiting out the commit-ack
-/// timeout. Errors before the header is complete carry no ack info.
-#[derive(Debug)]
-struct CrossFrameError {
-    message: String,
-    ack: Option<(Uuid, String, u64)>,
-}
-
-impl CrossFrameError {
-    fn new(message: impl Into<String>) -> Self {
-        Self {
-            message: message.into(),
-            ack: None,
-        }
-    }
-
-    fn with_ack(
-        message: impl Into<String>,
-        dataflow_id: Uuid,
-        shared_memory_id: String,
-        seq: u64,
-    ) -> Self {
-        Self {
-            message: message.into(),
-            ack: Some((dataflow_id, shared_memory_id, seq)),
-        }
-    }
-}
-
-/// Frame-parse + mirror-write core of the direct-TCP data plane (no zenoh
-/// involved), split out for unit testing. Returns the ack info
-/// `(dataflow id, pool id, seq)`; `Ok(None)` on clean EOF.
-async fn handle_cross_data_frame(
-    stream: &mut tokio::net::TcpStream,
-    tensor_pool: &TensorPoolManager,
-    machine_id: &str,
-) -> Result<Option<(Uuid, String, u64)>, CrossFrameError> {
-    use tokio::io::AsyncReadExt;
-
-    let mut magic = [0u8; 4];
-    match stream.read_exact(&mut magic).await {
-        Ok(_) => {}
-        Err(e) if e.kind() == std::io::ErrorKind::UnexpectedEof => return Ok(None),
-        Err(e) => return Err(CrossFrameError::new(format!("read magic: {e}"))),
-    }
-    if u32::from_be_bytes(magic) != CROSS_DATA_MAGIC {
-        return Err(CrossFrameError::new(
-            "bad frame magic (not a memory-pool data connection?)",
-        ));
-    }
-    let mut df_bytes = [0u8; 16];
-    stream
-        .read_exact(&mut df_bytes)
-        .await
-        .map_err(|e| CrossFrameError::new(format!("read dataflow id: {e}")))?;
-    let dataflow_id = Uuid::from_bytes(df_bytes);
-    let mut pool_len = [0u8; 4];
-    stream
-        .read_exact(&mut pool_len)
-        .await
-        .map_err(|e| CrossFrameError::new(format!("read pool id length: {e}")))?;
-    let pool_len = u32::from_be_bytes(pool_len) as usize;
-    if pool_len > 1024 {
-        return Err(CrossFrameError::new(format!(
-            "pool id too long ({pool_len} bytes)"
-        )));
-    }
-    let mut pool_bytes = vec![0u8; pool_len];
-    stream
-        .read_exact(&mut pool_bytes)
-        .await
-        .map_err(|e| CrossFrameError::new(format!("read pool id: {e}")))?;
-    let shared_memory_id =
-        String::from_utf8(pool_bytes).map_err(|_| CrossFrameError::new("pool id not UTF-8"))?;
-    let mut seq_bytes = [0u8; 8];
-    stream
-        .read_exact(&mut seq_bytes)
-        .await
-        .map_err(|e| CrossFrameError::new(format!("read seq: {e}")))?;
-    let seq = u64::from_be_bytes(seq_bytes);
-    let mut size_bytes = [0u8; 8];
-    stream
-        .read_exact(&mut size_bytes)
-        .await
-        .map_err(|e| CrossFrameError::new(format!("read size: {e}")))?;
-    let size = u64::from_be_bytes(size_bytes) as usize;
-
-    let dataflow_str = dataflow_id.to_string();
-    if !tensor_pool.is_cross(&dataflow_str, &shared_memory_id) {
-        return Err(CrossFrameError::with_ack(
-            format!("write for a pool without a cross-machine entry: {shared_memory_id}"),
-            dataflow_id,
-            shared_memory_id.clone(),
-            seq,
-        ));
-    }
-    let Some(shmem_name) =
-        TensorPoolManager::cross_pool_shmem_name(machine_id, &dataflow_str, &shared_memory_id)
-    else {
-        return Err(CrossFrameError::with_ack(
-            format!("invalid pool id {shared_memory_id}"),
-            dataflow_id,
-            shared_memory_id.clone(),
-            seq,
-        ));
-    };
-    // Serialise concurrent direct writes to the same pool of the same
-    // dataflow first (async lock: a std MutexGuard cannot be held across
-    // an await), so nothing non-Send crosses this await. Keyed by
-    // (dataflow, pool) — a bare pool id would serialize writes to
-    // different segments of concurrent dataflows.
-    let write_lock = {
-        let mut locks = CROSS_POOL_WRITE_LOCKS_ASYNC.lock().await;
-        match locks.get(&(dataflow_id, shared_memory_id.clone())) {
-            Some(lock) => lock.clone(),
-            None => {
-                let lock = std::sync::Arc::new(tokio::sync::Mutex::new(()));
-                locks.insert((dataflow_id, shared_memory_id.clone()), lock.clone());
-                lock
-            }
-        }
-    };
-    let _guard = write_lock.lock().await;
-    // Open + validate the mirror (all sync, no awaits while raw pointers
-    // are live), then read the payload straight into the data region
-    // under the seqlock — zero user-space copies on this side. The open
-    // lives inside a block so the `Shmem` local (with its drop flag) is
-    // consumed before the read await.
-    let mut writer = {
-        let shmem = ShmemConf::new().os_id(&shmem_name).open().map_err(|e| {
-            CrossFrameError::with_ack(
-                format!("cannot open mirror {shmem_name}: {e}"),
-                dataflow_id,
-                shared_memory_id.clone(),
-                seq,
-            )
-        })?;
-        let shmem_ptr = shmem.as_ptr();
-        let magic8 = unsafe { std::slice::from_raw_parts(shmem_ptr, 8) };
-        if magic8 != DORADMA_MAGIC {
-            return Err(CrossFrameError::with_ack(
-                format!("{shared_memory_id} header magic mismatch"),
-                dataflow_id,
-                shared_memory_id.clone(),
-                seq,
-            ));
-        }
-        // The frame `size` must equal the registered pool size recorded
-        // in the mirror header: a short frame would otherwise publish an
-        // even generation over a truncated tensor that readers accept as
-        // complete (the segment-length bound alone cannot catch it, since
-        // the segment is created at the registered size + header).
-        let registered_size = {
-            let json_len = unsafe { read_header_u64(shmem_ptr.add(8)) } as usize;
-            let json_bytes =
-                unsafe { std::slice::from_raw_parts(shmem_ptr.add(DORADMA_HEADER_SIZE), json_len) };
-            serde_json::from_slice::<serde_json::Value>(json_bytes)
-                .ok()
-                .and_then(|v| v.get("size").and_then(|s| s.as_u64()))
-                .unwrap_or(0) as usize
-        };
-        if size != registered_size {
-            return Err(CrossFrameError::with_ack(
-                format!(
-                    "{shared_memory_id} frame size {size} != registered pool size {registered_size}"
-                ),
-                dataflow_id,
-                shared_memory_id.clone(),
-                seq,
-            ));
-        }
-        let data_offset = unsafe { read_header_u64(shmem_ptr.add(16)) } as usize;
-        // Checked add: `size` is wire-controlled (u64 read straight off
-        // the socket), so `data_offset + size` can wrap to a small value
-        // and pass the bounds check — then a `size`-byte slice would be
-        // constructed past the mapping (UB; a remote-triggerable abort in
-        // debug builds). Reject the overflow explicitly.
-        let Some(end) = data_offset.checked_add(size) else {
-            return Err(CrossFrameError::with_ack(
-                format!("{shared_memory_id} data_offset {data_offset} + {size} overflows usize"),
-                dataflow_id,
-                shared_memory_id.clone(),
-                seq,
-            ));
-        };
-        if end > shmem.len() {
-            return Err(CrossFrameError::with_ack(
-                format!(
-                    "{shared_memory_id} data_offset {data_offset} + {size} exceeds shmem size {}",
-                    shmem.len()
-                ),
-                dataflow_id,
-                shared_memory_id.clone(),
-                seq,
-            ));
-        }
-        DirectMirrorWriter::new(shmem, data_offset, size)
-    };
-    let dst = writer.data_slice_mut();
-    if let Err(e) = stream.read_exact(dst).await {
-        // The writer drops without `finish`: the seqlock generation stays
-        // odd (in-progress) — readers reject the torn frame and the next
-        // full write self-heals (see the NOTE on `DirectMirrorWriter`).
-        // The ack info lets the caller fail the origin's pending write
-        // instead of stranding it for the timeout.
-        return Err(CrossFrameError::with_ack(
-            format!("read payload: {e}"),
-            dataflow_id,
-            shared_memory_id,
-            seq,
-        ));
-    }
-    writer.finish();
-    Ok(Some((dataflow_id, shared_memory_id, seq)))
-}
-
-/// Send one direct-TCP data frame to a peer's data listener (origin side).
-/// Reuses a persistent connection per endpoint; a dead connection is
-/// dropped and re-established lazily. The connection is taken out of the
-/// map while in flight (a std `MutexGuard` cannot be held across an
-/// `.await`), so concurrent writers to the same endpoint serialize on the
-/// map lock instead — fine for the turn-based benchmark cadence. When
-/// auth is configured, a freshly connected socket performs the token
-/// handshake before the first frame.
-async fn send_cross_data_frame(
-    conns: &Arc<std::sync::Mutex<HashMap<std::net::SocketAddr, tokio::net::TcpStream>>>,
-    endpoint: std::net::SocketAddr,
-    dataflow_id: Uuid,
-    shared_memory_id: &str,
-    seq: u64,
-    data: &[u8],
-) -> Result<(), String> {
-    use tokio::io::AsyncWriteExt;
-
-    let mut stream = {
-        // The guard is a temporary: it must be dropped before the connect
-        // await below (a std MutexGuard is not Send across awaits).
-        let existing = conns
-            .lock()
-            .unwrap_or_else(|e| e.into_inner())
-            .remove(&endpoint);
-        match existing {
-            Some(stream) => stream,
-            None => {
-                let stream = tokio::time::timeout(
-                    std::time::Duration::from_secs(5),
-                    tokio::net::TcpStream::connect(endpoint),
-                )
-                .await
-                .map_err(|_| format!("connect timeout to {endpoint}"))?
-                .map_err(|e| format!("connect to {endpoint} failed: {e}"))?;
-                // Auth handshake on fresh connections only. A rejected
-                // handshake is an error on this side too — the caller
-                // degrades to the zenoh relay.
-                let mut stream = stream;
-                if let Some(token) = cross_data_auth_token() {
-                    tokio::time::timeout(
-                        CROSS_DATA_READ_TIMEOUT,
-                        auth_handshake_send(&mut stream, &token),
-                    )
-                    .await
-                    .map_err(|_| format!("auth handshake timeout to {endpoint}"))?
-                    .map_err(|e| format!("auth handshake to {endpoint} failed: {e}"))?;
-                }
-                stream
-            }
-        }
-    };
-    let mut buf = Vec::with_capacity(4 + 16 + 4 + shared_memory_id.len() + 8 + 8 + data.len());
-    buf.extend_from_slice(&CROSS_DATA_MAGIC.to_be_bytes());
-    buf.extend_from_slice(dataflow_id.as_bytes());
-    buf.extend_from_slice(&(shared_memory_id.len() as u32).to_be_bytes());
-    buf.extend_from_slice(shared_memory_id.as_bytes());
-    buf.extend_from_slice(&seq.to_be_bytes());
-    buf.extend_from_slice(&(data.len() as u64).to_be_bytes());
-    let result = async {
-        stream.write_all(&buf).await?;
-        stream.write_all(data).await?;
-        stream.flush().await?;
-        Ok::<(), std::io::Error>(())
-    }
-    .await;
-    if let Err(e) = result {
-        // Dead connection — drop it (not re-inserted) so the next write
-        // reconnects.
-        return Err(format!("direct write to {endpoint} failed: {e}"));
-    }
-    conns
-        .lock()
-        .unwrap_or_else(|e| e.into_inner())
-        .insert(endpoint, stream);
-    Ok(())
-}
-
-/// Read a pool's tensor bytes from its local segment — the
-/// shared-memory-reference write path: the node's write request carries
-/// only `(id, size)` metadata, so the daemon opens the sender's segment
-/// (name recorded at registration) and copies `size` bytes from the
-/// DORADMA data region. Keeping the node→daemon request KB-scale removes
-/// the transport cap on cross-machine pool size (the request previously
-/// carried the whole tensor, bounded by `dora_message::MAX_MESSAGE_BYTES`).
-fn read_pool_segment_data(shmem_name: &str, size: usize) -> Result<Vec<u8>, String> {
-    let shmem = ShmemConf::new()
-        .os_id(shmem_name)
-        .open()
-        .map_err(|e| format!("cannot open segment {shmem_name}: {e}"))?;
-    let shmem_ptr = shmem.as_ptr();
-    // Guard against a corrupt/truncated header before any pointer math.
-    let magic = unsafe { std::slice::from_raw_parts(shmem_ptr, 8) };
-    if magic != DORADMA_MAGIC {
-        return Err(format!("segment {shmem_name} header magic mismatch"));
-    }
-    let data_offset = unsafe { read_header_u64(shmem_ptr.add(16)) } as usize;
-    // Checked add: same wrapping concern as the direct-TCP frame path
-    // (size comes from the node request, not the socket, but a corrupt
-    // header's data_offset must not wrap the bounds check either).
-    let Some(end) = data_offset.checked_add(size) else {
-        return Err(format!(
-            "segment {shmem_name} data_offset {data_offset} + {size} overflows usize"
-        ));
-    };
-    if end > shmem.len() {
-        return Err(format!(
-            "segment {shmem_name} data_offset {data_offset} + {size} exceeds shmem size {}",
-            shmem.len()
-        ));
-    }
-    let mut data = vec![0u8; size];
-    unsafe {
-        std::ptr::copy_nonoverlapping(shmem_ptr.add(data_offset), data.as_mut_ptr(), size);
-    }
-    Ok(data)
-}
-
-/// Remove a mirrored cross-machine pool's shmem segment. Linux keeps
-/// pools in /dev/shm; the name is only removable by file unlink because
-/// the mirror handle was dropped owner-less (`set_owner(false)`).
-/// Path-traversal guarded, mirroring the memory-pool crate's
-/// `free_shared_memory` checks.
-fn remove_cross_pool_shmem(shmem_name: &str) {
-    if !shmem_name.starts_with("dora_pool_")
-        || shmem_name.contains('/')
-        || shmem_name.contains("..")
-    {
-        tracing::warn!("memory pool: refusing to remove shmem `{shmem_name}`: unexpected name");
-        return;
-    }
-    #[cfg(target_os = "linux")]
-    {
-        let shm_path = format!("/dev/shm/{shmem_name}");
-        match std::fs::remove_file(&shm_path) {
-            Ok(_) => {}
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => {}
-            Err(e) => {
-                tracing::warn!(
-                    "memory pool: failed to unlink shared memory file {}: {}. \
-                     The file may still be in use by other processes.",
-                    shm_path,
-                    e
-                );
-            }
-        }
-    }
-    #[cfg(not(target_os = "linux"))]
-    {
-        // Not an anomaly worth a warning per pool per free/finish: the
-        // whole cross-machine mirror path is Linux-only.
-        tracing::debug!("memory pool: shmem removal only implemented on Linux");
-    }
-}
-
-/// Unlink the `/dev/shm` segment a cross-machine pool resolves to under
-/// `machine_id`, returning whether the name could be derived.
-///
-/// Used by the free paths that only know the pool id. Paths that hold
-/// the name the segment was created under should unlink that instead —
-/// see `cross_pool_shmem_name`'s own warning: "create, write, and the
-/// free paths must agree or a mirror leaks under a name nobody unlinks."
-fn unlink_cross_pool_segment(machine_id: &str, dataflow_id: &str, shared_memory_id: &str) -> bool {
-    let Some(shmem_name) =
-        TensorPoolManager::cross_pool_shmem_name(machine_id, dataflow_id, shared_memory_id)
-    else {
-        tracing::warn!("memory pool: invalid pool id {shared_memory_id}, cannot unlink mirror");
-        return false;
-    };
-    remove_cross_pool_shmem(&shmem_name);
-    true
-}
-
-/// Publish an inter-daemon memory-pool event over the dataflow topic.
-/// serialize + declare + put all run off the event loop (Block congestion
-/// control can block declare_publisher on a degraded link). Logs the
-/// declare and put timing — postcard::serialize of the 61.44MB
-/// WriteMemoryPool payload takes hundreds of ms (3s+ in debug builds),
-/// so the timing is worth knowing on every publish path.
-async fn publish_memory_pool_event(
-    session: &zenoh::Session,
-    clock: &Arc<HLC>,
-    dataflow_id: &Uuid,
-    event: &InterDaemonEvent,
-    shm_provider: Option<&ShmProvider<PosixShmProviderBackend>>,
-) -> eyre::Result<()> {
-    let serialized = Timestamped {
-        inner: event.clone(),
-        timestamp: clock.new_timestamp(),
-    }
-    .serialize()?;
-    let payload_len = serialized.len();
-    let topic = dataflow_memory_pool_topic(dataflow_id);
-    // Zenoh errors are boxed trait objects — eyre's `From` conversion
-    // needs a Sized error, so convert explicitly instead of `?`.
-    let declared = std::time::Instant::now();
-    let publisher = session
-        .declare_publisher(topic.clone())
-        .congestion_control(CongestionControl::Block)
-        // Remote-only: with the default Locality::Any the publisher's own
-        // subscriber receives its own put, and on the RegisterPoolAck path
-        // that local echo would be a self-ack that races (and beats) the
-        // remote ack (see the RegisterPool handler).
-        .allowed_destination(Locality::Remote)
-        .await
-        .map_err(|e| eyre!("memory pool: declare_publisher({topic}) failed: {e}"))?;
-    tracing::info!(
-        "memory pool: declared {topic} in {:?}, starting put ({payload_len} bytes)",
-        declared.elapsed()
-    );
-    let started = std::time::Instant::now();
-    // Control events (RegisterPool/RegisterPoolAck/FreePool) go over zenoh
-    // SHM when a provider exists: same-host daemons map the payload
-    // zero-copy, cross-host receivers get an implicit copy from the zenoh
-    // transport. MemoryPoolWrite carries the cross-machine tensor data and
-    // only ever happens cross-host, where an SHM segment would be an extra
-    // copy with no benefit — keep it on the plain path.
-    let put_result = if matches!(event, InterDaemonEvent::MemoryPoolWrite { .. }) {
-        publisher.put(serialized).await
-    } else if let Some(provider) = shm_provider {
-        // Synchronous wait: control payloads are KB-scale, so the shm
-        // segment allocation is microseconds — no need for the async
-        // allocation policy machinery.
-        match provider.alloc(payload_len).wait() {
-            // `alloc` guarantees a buffer of at least `payload_len` bytes
-            // (alignment may round up), so the copy cannot overflow.
-            Ok(mut buf) => {
-                let buf_slice: &mut [u8] = buf.as_mut();
-                buf_slice[..payload_len].copy_from_slice(&serialized);
-                let payload: ZBytes = buf.into();
-                publisher.put(payload).await
-            }
-            Err(e) => {
-                tracing::warn!(
-                    "memory pool: SHM alloc failed ({e}), falling back to regular payload"
-                );
-                publisher.put(serialized).await
-            }
-        }
-    } else {
-        publisher.put(serialized).await
-    };
-    put_result.map_err(|e| eyre!("memory pool: publish to {topic} failed: {e}"))?;
-    tracing::info!(
-        "memory pool: put to {topic} completed in {:?}",
-        started.elapsed()
-    );
-    Ok(())
-}
-
-/// Release a cross-machine pool from the freeing daemon: unlink this
-/// machine's mirror (self-machine-qualified name; on the origin daemon
-/// the unlink is a harmless NotFound no-op) and publish a targeted
-/// `FreePool` so the peer drops its tracking entry. The publish is
-/// Remote-only — the initiator never receives its own echo, so it must
-/// unlink its own mirror here. The caller has already removed the
-/// cross_pools entry and passes the recorded peer (the pool's other
-/// machine) as the free target.
-async fn release_cross_pool(
-    session: &zenoh::Session,
-    clock: &Arc<HLC>,
-    dataflow_id: &Uuid,
-    machine_id: &str,
-    peer_machine_id: &str,
-    shared_memory_id: &str,
-    shm_provider: Option<&ShmProvider<PosixShmProviderBackend>>,
-) {
-    if !unlink_cross_pool_segment(machine_id, &dataflow_id.to_string(), shared_memory_id) {
-        return;
-    }
-    tracing::info!("memory pool: forwarding free of {shared_memory_id} to peer {peer_machine_id}");
-    if let Err(e) = publish_memory_pool_event(
-        session,
-        clock,
-        dataflow_id,
-        &InterDaemonEvent::FreePool {
-            dataflow_id: *dataflow_id,
-            machine_id: peer_machine_id.to_string(),
-            shared_memory_id: shared_memory_id.to_string(),
-        },
-        shm_provider,
-    )
-    .await
-    {
-        tracing::warn!("memory pool: failed to publish FreePool for {shared_memory_id}: {e}");
-    }
-}
-
-/// Pending synchronous register confirmations:
-/// (dataflow id, pool id) -> ack channel. Keyed by dataflow too: every
-/// node process restarts its pool counter from zero, so a bare pool id
-/// repeats across concurrently running dataflows and an ack could
-/// satisfy the wrong registration.
-type RegisterAckSenders = std::sync::Mutex<
-    std::collections::HashMap<
-        (Uuid, String),
-        tokio::sync::oneshot::Sender<(bool, bool, Option<u16>, Option<std::net::SocketAddr>)>,
-    >,
->;
-static CROSS_REGISTER_PENDING: std::sync::LazyLock<RegisterAckSenders> =
-    std::sync::LazyLock::new(RegisterAckSenders::default);
-
-/// Pending cross-machine write replies:
-/// (dataflow id, pool id, write seq) -> the node's reply channel. The
-/// write reply is withheld until the mirror daemon confirms the segment
-/// write (`MemoryPoolWriteAck`), so the output notification that follows
-/// the write can never overtake the tensor data.
-type CrossWriteReplySenders = std::sync::Mutex<
-    std::collections::HashMap<(Uuid, String, u64), tokio::sync::oneshot::Sender<DaemonReply>>,
->;
-static CROSS_WRITE_PENDING: std::sync::LazyLock<CrossWriteReplySenders> =
-    std::sync::LazyLock::new(CrossWriteReplySenders::default);
-
-/// Per-pool write sequence counters: (dataflow id, pool id) -> next seq.
-/// Assigned at the origin, echoed by the mirror's commit ack, so the
-/// ack can never resolve a reply for a different write.
-static CROSS_WRITE_SEQ: std::sync::LazyLock<
-    std::sync::Mutex<std::collections::HashMap<(Uuid, String), u64>>,
-> = std::sync::LazyLock::new(std::sync::Mutex::default);
-
-/// Pools whose direct-TCP write path is currently degraded to the zenoh
-/// relay. The fallback is the steady state on a broken link, so without
-/// this tracking every frame would warn — flooding the log. Keyed like
-/// [`CROSS_WRITE_PENDING`]: `(dataflow id, pool id)`. Drained per
-/// dataflow in `finish_dataflow`.
-static CROSS_DIRECT_DEGRADED: std::sync::LazyLock<
-    std::sync::Mutex<std::collections::HashSet<(Uuid, String)>>,
-> = std::sync::LazyLock::new(std::sync::Mutex::default);
-
-/// Mark a pool's direct-TCP path as degraded. Returns `true` only on the
-/// **first** degradation — the caller warns exactly then; repeated
-/// failures while already degraded stay silent.
-fn note_direct_degraded(dataflow_id: Uuid, shared_memory_id: &str) -> bool {
-    CROSS_DIRECT_DEGRADED
-        .lock()
-        .unwrap_or_else(|e| e.into_inner())
-        .insert((dataflow_id, shared_memory_id.to_string()))
-}
-
-/// Mark a pool's direct-TCP path as recovered. Returns `true` only if the
-/// pool was actually degraded — the caller logs the recovery exactly
-/// once, on the first successful direct write after a fallback.
-fn note_direct_recovered(dataflow_id: Uuid, shared_memory_id: &str) -> bool {
-    CROSS_DIRECT_DEGRADED
-        .lock()
-        .unwrap_or_else(|e| e.into_inner())
-        .remove(&(dataflow_id, shared_memory_id.to_string()))
-}
-
-/// How long a cross-machine write waits for the remote commit ack before
-/// failing loudly. Generous: the WAN transfer of a near-limit frame alone
-/// can take tens of seconds; a dead link fails earlier via the publish
-/// error path.
-const CROSS_WRITE_ACK_TIMEOUT: std::time::Duration = std::time::Duration::from_secs(120);
-/// Upper bound for reading one direct-TCP data frame (header + payload).
-/// A peer that sends a header with a large `size` and then stalls would
-/// otherwise hold the per-pool lock and leave the seqlock odd
-/// indefinitely, wedging every subsequent write to that pool. 300s covers
-/// a 1 GiB frame on a ~5 MB/s slow WAN link; on timeout the connection is
-/// dropped, the lock guard drops, and the odd generation marks the frame
-/// torn (next write self-heals).
-const CROSS_DATA_READ_TIMEOUT: std::time::Duration = std::time::Duration::from_secs(300);
-
-/// Resolve the pending cross-machine write reply for a commit ack.
-/// Only the seq-matched pending entry is resolved — an ack for a previous
-/// write can never satisfy a newer pending reply. Returns whether a
-/// matching pending entry existed and was resolved.
-fn resolve_cross_write_ack(
-    dataflow_id: Uuid,
-    shared_memory_id: String,
-    seq: u64,
-    ok: bool,
-    error: Option<String>,
-) -> bool {
-    if let Some(tx) = CROSS_WRITE_PENDING
-        .lock()
-        .unwrap_or_else(|e| e.into_inner())
-        .remove(&(dataflow_id, shared_memory_id, seq))
-    {
-        let result = if ok {
-            Ok(())
-        } else {
-            Err(error.unwrap_or_else(|| "remote mirror write failed".to_string()))
-        };
-        let _ = tx.send(DaemonReply::Result(result));
-        true
-    } else {
-        false
-    }
-}
-
-/// Fire the safety-net timeout for a pending cross-machine write: remove the
-/// entry under the write's *effective* seq and fail the node's blocked
-/// `WritePinnedMemory`. Returns whether an entry existed.
-///
-/// `effective_seq` is the seq the entry is live under now, which advances from
-/// the original `seq` to a fresh `relay_seq` on a direct-TCP → zenoh failover.
-/// Keying the timeout to the pre-failover seq makes it a no-op after a re-key,
-/// leaving the write hung forever (#3193).
-///
-/// The load happens *under the pending-map guard*, the same one
-/// `rekey_cross_write_to_relay` holds while it re-keys and stores. Loading
-/// outside it would reopen #3193 as a race: a failover landing between the load
-/// and the `remove` would leave the removal targeting a key that has already
-/// moved, and nothing retries.
-fn fire_cross_write_timeout(
-    dataflow_id: Uuid,
-    shared_memory_id: &str,
-    effective_seq: &std::sync::atomic::AtomicU64,
-) -> bool {
-    let mut pending = CROSS_WRITE_PENDING
-        .lock()
-        .unwrap_or_else(|e| e.into_inner());
-    let seq = effective_seq.load(std::sync::atomic::Ordering::SeqCst);
-    let Some(tx) = pending.remove(&(dataflow_id, shared_memory_id.to_string(), seq)) else {
-        return false;
-    };
-    drop(pending);
-    let _ = tx.send(DaemonReply::Result(Err(
-        "cross-machine write timed out waiting for the remote commit ack".to_string(),
-    )));
-    true
-}
-
-/// Re-key a pending cross-machine write from `seq` to a fresh relay seq for the
-/// zenoh fallback hop, advancing `effective_seq` to match. Returns the new
-/// relay seq, or `None` if nothing was pending under `seq` (already resolved).
-///
-/// The re-key drops the stale direct-plane `ok:false` (published the instant a
-/// mid-frame direct read fails) via the first-wins resolver, and keeps the
-/// per-pool counter monotonic so the next write cannot collide with the
-/// re-keyed seq. Advancing `effective_seq` is the crucial half: the safety-net
-/// timeout task reads it, so a relay ack lost after the re-key is still
-/// reclaimed rather than hanging the node forever (#3193).
-fn rekey_cross_write_to_relay(
-    dataflow_id: Uuid,
-    shared_memory_id: &str,
-    seq: u64,
-    effective_seq: &std::sync::atomic::AtomicU64,
-) -> Option<u64> {
-    // Remove, re-key and store under a *single* pending-map guard. Releasing
-    // it between the remove and the insert would leave the entry in no map at
-    // all, so a concurrent `drain_cross_write_pending` walks past it and the
-    // re-insert then outlives `finish_dataflow` — the node stays blocked until
-    // the 120s safety net and the entry leaks past the dataflow it belongs to.
-    // The counter bump nests `CROSS_WRITE_SEQ` inside this guard; that is the
-    // only nesting of the two (every other site drops the seq lock before
-    // touching the pending map), so the order cannot deadlock.
-    let mut pending = CROSS_WRITE_PENDING
-        .lock()
-        .unwrap_or_else(|e| e.into_inner());
-    let tx = pending.remove(&(dataflow_id, shared_memory_id.to_string(), seq))?;
-    let relay_seq = {
-        let mut seqs = CROSS_WRITE_SEQ.lock().unwrap_or_else(|e| e.into_inner());
-        let counter = seqs
-            .entry((dataflow_id, shared_memory_id.to_string()))
-            .or_insert(0);
-        *counter += 1;
-        *counter
-    };
-    pending.insert((dataflow_id, shared_memory_id.to_string(), relay_seq), tx);
-    // Still under the guard: `fire_cross_write_timeout` loads this while
-    // holding the same lock, so it can never observe a stale seq (#3193).
-    effective_seq.store(relay_seq, std::sync::atomic::Ordering::SeqCst);
-    Some(relay_seq)
-}
-
-/// Fail and drop every cross-machine write reply still pending for
-/// `dataflow_id`, unblocking any node stuck in `WritePinnedMemory`. Returns the
-/// number of stranded writes reclaimed.
-///
-/// Called from `finish_dataflow`: the per-write safety-net timeout normally
-/// reclaims these, but a dataflow that finishes first (node crash, `dora stop`,
-/// or a relay entry orphaned by a failover) would otherwise leave the entry —
-/// and the node's blocked write — hanging until that timeout, and leak the
-/// entry for the daemon's lifetime if the reply channel is already gone (#3193).
-/// Mirrors how the sibling cross-write maps are already drained here.
-fn drain_cross_write_pending(dataflow_id: Uuid) -> usize {
-    let stranded: Vec<_> = {
-        let mut pending = CROSS_WRITE_PENDING
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-        // One pass, no key clones: the previous scan-clone-then-remove shape
-        // allocated a `String` per match just to look each entry back up.
-        pending
-            .extract_if(|(df, _, _), _| *df == dataflow_id)
-            .map(|(_, tx)| tx)
-            .collect()
-    };
-    let count = stranded.len();
-    for tx in stranded {
-        let _ = tx.send(DaemonReply::Result(Err(
-            "dataflow finished before the cross-machine write was acknowledged".to_string(),
-        )));
-    }
-    count
-}
-
-/// Size of the daemon's zenoh SHM provider segment. Memory-pool control
-/// notifications (RegisterPool/RegisterPoolAck/FreePool) are KB-scale,
-/// so a small segment carries all in-flight control traffic with headroom;
-/// the cross-machine tensor payload (MemoryPoolWrite) deliberately stays
-/// on the plain path and never allocates from here.
-const MEMORY_POOL_SHM_PROVIDER_SIZE: usize = 8 * 1024 * 1024;
+mod memory_pool;
+use crate::memory_pool::*;
 
 /// A dataflow's memory-pool zenoh subscriber, plus the liveness flag the
 /// detached mirror-creation tasks it feeds check before they publish a
@@ -10431,320 +8983,6 @@ fn rebuild_debug_topic_stream(
 }
 
 #[cfg(test)]
-mod cross_mirror_descriptor_tests {
-    use super::build_mirror_descriptor;
-    use dora_message::metadata::{MetadataParameters, Parameter};
-
-    fn decode(bytes: &[u8]) -> MetadataParameters {
-        serde_json::from_slice(bytes).expect("descriptor must decode")
-    }
-
-    fn get_str<'a>(params: &'a MetadataParameters, key: &str) -> &'a str {
-        match params.get(key) {
-            Some(Parameter::String(s)) => s,
-            other => panic!("expected string {key}, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn gpu_mirror_descriptor_carries_receiver_device() {
-        // A GPU receiver's mirror: pinned_type must be the receiver
-        // device ("cuda:0"), not "cpu" — the python receiver derives
-        // `effective_as_cuda` from it and stages the mirror's CPU data
-        // region HtoD.
-        let bytes = build_mirror_descriptor(
-            61_440_000,
-            "torch.int64",
-            &[7_680_000],
-            "dora_pool_D_01a00445-b12d-734b-852b-463d76c749b1_sender_node_1",
-            "cuda:0",
-        );
-        let params = decode(&bytes);
-        assert_eq!(
-            get_str(&params, "pinned_type"),
-            "cuda:0",
-            "GPU mirror must advertise the receiver device"
-        );
-        assert_eq!(
-            get_str(&params, "dtype"),
-            "torch.int64",
-            "dtype must round-trip"
-        );
-        assert_eq!(
-            params.get("size"),
-            Some(&Parameter::Integer(61_440_000)),
-            "size must round-trip"
-        );
-        assert_eq!(
-            params.get("shape"),
-            Some(&Parameter::ListInt(vec![7_680_000])),
-            "shape must round-trip"
-        );
-        assert_eq!(
-            get_str(&params, "shared_memory_name",),
-            "dora_pool_D_01a00445-b12d-734b-852b-463d76c749b1_sender_node_1",
-            "shared_memory_name must be the machine-qualified mirror segment"
-        );
-        assert_eq!(
-            params.get("ipc_present"),
-            Some(&Parameter::Bool(false)),
-            "cross-machine mirrors never export a host-local IPC handle"
-        );
-    }
-
-    #[test]
-    fn cpu_mirror_descriptor_is_cpu() {
-        // A CPU receiver's mirror: pinned_type "cpu" keeps the python
-        // receiver on the CPU read path (no HtoD staging).
-        let bytes = build_mirror_descriptor(
-            61_440_000,
-            "torch.int64",
-            &[7_680_000],
-            "dora_pool_D_01a00445-b12d-734b-852b-463d76c749b1_sender_node_1",
-            "cpu",
-        );
-        assert_eq!(get_str(&decode(&bytes), "pinned_type"), "cpu");
-    }
-}
-
-#[cfg(test)]
-mod cross_write_failover_tests {
-    use super::{
-        CROSS_WRITE_PENDING, CROSS_WRITE_SEQ, drain_cross_write_pending, fire_cross_write_timeout,
-        rekey_cross_write_to_relay, resolve_cross_write_ack,
-    };
-    use dora_message::DataflowId;
-    use dora_message::daemon_to_node::DaemonReply;
-
-    #[test]
-    fn failover_rekey_drops_stale_direct_ack_and_resolves_via_relay() {
-        // The WriteMemoryPool failover rekeys the pending entry to a
-        // fresh seq for the zenoh relay hop, so the mirror's ok:false
-        // for the abandoned direct seq (published the instant its
-        // mid-frame read fails) is dropped by the first-wins resolver,
-        // and the relay's ok:true resolves the node's write. Without the
-        // rekey, the ok:false would win the race and fail a write the
-        // relay delivered byte-for-byte (bot review 5307022693).
-        let df = DataflowId::new_v4();
-        let pool = "pool_sender_node_1".to_string();
-        let (tx, rx) = tokio::sync::oneshot::channel();
-        // Write-side seq allocation (counter 0 -> 1), mirroring the
-        // WriteMemoryPool handler.
-        let mut seqs = CROSS_WRITE_SEQ.lock().unwrap_or_else(|e| e.into_inner());
-        let counter = seqs.entry((df, pool.clone())).or_insert(0);
-        *counter += 1;
-        let seq = *counter;
-        drop(seqs);
-        assert_eq!(seq, 1, "test setup assumes the first write takes seq 1");
-        CROSS_WRITE_PENDING
-            .lock()
-            .unwrap_or_else(|e| e.into_inner())
-            .insert((df, pool.clone(), seq), tx);
-
-        // Rekey via the production helper the direct-failure arm calls, so
-        // this asserts against the real path rather than a copy of it.
-        let effective_seq = std::sync::atomic::AtomicU64::new(seq);
-        let relay_seq = rekey_cross_write_to_relay(df, &pool, seq, &effective_seq)
-            .expect("pending entry present");
-        assert_ne!(relay_seq, seq, "rekey must produce a fresh seq");
-
-        // The stale direct-plane ok:false arrives first: no pending entry
-        // under the old seq, so it is dropped without touching the node.
-        assert!(!resolve_cross_write_ack(
-            df,
-            pool.clone(),
-            seq,
-            false,
-            Some("mid-frame direct read failed".to_string())
-        ));
-
-        // The relay's ok:true resolves the node's write as a success.
-        assert!(resolve_cross_write_ack(
-            df,
-            pool.clone(),
-            relay_seq,
-            true,
-            None
-        ));
-        let result = rx.blocking_recv().expect("node reply delivered");
-        assert!(
-            matches!(result, DaemonReply::Result(Ok(()))),
-            "node must see the relay's success, not the stale direct failure"
-        );
-    }
-
-    #[test]
-    fn rekey_advances_effective_seq_so_the_timeout_follows_it() {
-        // #3193: drives the production re-key path (`rekey_cross_write_to_relay`,
-        // the direct-TCP → zenoh failover arm) and pins the actual fix — that
-        // it advances the shared `effective_seq` the safety-net timeout reads.
-        // Deleting the `effective_seq.store(relay_seq)` inside the helper makes
-        // the `effective_seq` assertion below fail, and the timeout then reads
-        // the stale seq (a no-op) so the final resolve fails too — i.e. this
-        // test fails exactly when the write would hang forever.
-        let df = DataflowId::new_v4();
-        let pool = "pool_rekey_atomic".to_string();
-        let (tx, mut rx) = tokio::sync::oneshot::channel();
-
-        // Allocate the seq exactly as the handler does (counter 0 -> 1) and
-        // insert the pending reply under it.
-        let seq = {
-            let mut seqs = CROSS_WRITE_SEQ.lock().unwrap_or_else(|e| e.into_inner());
-            let counter = seqs.entry((df, pool.clone())).or_insert(0);
-            *counter += 1;
-            *counter
-        };
-        CROSS_WRITE_PENDING
-            .lock()
-            .unwrap_or_else(|e| e.into_inner())
-            .insert((df, pool.clone(), seq), tx);
-        let effective_seq = std::sync::atomic::AtomicU64::new(seq);
-
-        // Run the real re-key the handler runs on a direct-send failure.
-        let relay_seq = rekey_cross_write_to_relay(df, &pool, seq, &effective_seq)
-            .expect("a pending entry under the original seq");
-        assert_ne!(relay_seq, seq, "re-key must allocate a fresh monotonic seq");
-        // The crux of the fix: the atomic the timeout task reads was advanced
-        // to the re-keyed seq.
-        assert_eq!(
-            effective_seq.load(std::sync::atomic::Ordering::SeqCst),
-            relay_seq,
-            "re-key must advance the effective seq the timeout follows"
-        );
-
-        // A timeout still keyed to the pre-failover seq is a no-op: the entry
-        // moved. This is the #3193 regression, reproduced by handing the
-        // helper an atomic frozen at the stale seq.
-        let stale_seq = std::sync::atomic::AtomicU64::new(seq);
-        assert!(!fire_cross_write_timeout(df, &pool, &stale_seq));
-        assert!(
-            rx.try_recv().is_err(),
-            "write must still be pending after a timeout keyed to the stale seq"
-        );
-
-        // The real safety net passes the shared atomic the re-key advanced —
-        // the helper loads it under the pending guard — so it resolves the
-        // re-keyed entry and fails the node's write cleanly instead of hanging.
-        assert!(fire_cross_write_timeout(df, &pool, &effective_seq));
-        assert!(
-            matches!(rx.try_recv(), Ok(DaemonReply::Result(Err(_)))),
-            "the timeout keyed to the effective seq must fail the node's write"
-        );
-    }
-
-    #[test]
-    fn finish_dataflow_drain_fails_pending_cross_writes() {
-        // #3193: a dataflow that finishes with a write still pending (node
-        // crash / stop, or a relay entry orphaned by failover) must have that
-        // entry reclaimed and the node unblocked, not left to leak. An
-        // unrelated dataflow's pending write must survive the drain.
-        let df = DataflowId::new_v4();
-        let other_df = DataflowId::new_v4();
-        let pool = "pool_finish_drain".to_string();
-        let (tx, mut rx) = tokio::sync::oneshot::channel();
-        let (other_tx, mut other_rx) = tokio::sync::oneshot::channel();
-
-        {
-            let mut pending = CROSS_WRITE_PENDING
-                .lock()
-                .unwrap_or_else(|e| e.into_inner());
-            pending.insert((df, pool.clone(), 7), tx);
-            pending.insert((other_df, pool.clone(), 7), other_tx);
-        }
-
-        assert_eq!(drain_cross_write_pending(df), 1);
-        assert!(
-            matches!(rx.try_recv(), Ok(DaemonReply::Result(Err(_)))),
-            "the finishing dataflow's pending write must be failed"
-        );
-        assert!(
-            other_rx.try_recv().is_err(),
-            "an unrelated dataflow's pending write must be untouched"
-        );
-
-        // Leave the shared static tidy for other tests.
-        drain_cross_write_pending(other_df);
-    }
-
-    #[cfg(target_os = "linux")]
-    #[test]
-    fn create_cross_pool_shmem_idempotent_on_matching_size() {
-        // A retried registration whose first ack was lost re-runs
-        // create_cross_pool_shmem against the daemon's own live mirror:
-        // same (dataflow, pool) and same size must be an idempotent
-        // success, not an unlink+recreate that strands receivers already
-        // mapping the old inode (bot review 5307022693).
-        let df = DataflowId::new_v4();
-        let pool = "pool_sender_node_1".to_string();
-        let size = 61_440_000usize;
-        super::create_cross_pool_shmem(
-            &df,
-            "M",
-            &pool,
-            size,
-            "torch.int64",
-            &[7_680_000],
-            "cpu",
-            None,
-        )
-        .expect("first create");
-        let name = super::TensorPoolManager::cross_pool_shmem_name("M", &df.to_string(), &pool)
-            .expect("name");
-        let path = format!("/dev/shm/{name}");
-        let ino_after_create = {
-            use std::os::unix::fs::MetadataExt;
-            std::fs::metadata(&path).expect("segment exists").ino()
-        };
-        // Second attempt (same everything): idempotent success — the
-        // inode must survive, so receivers mapping it keep reading the
-        // live mirror rather than a frozen orphan.
-        super::create_cross_pool_shmem(
-            &df,
-            "M",
-            &pool,
-            size,
-            "torch.int64",
-            &[7_680_000],
-            "cpu",
-            None,
-        )
-        .expect("retry must be idempotent");
-        let ino_after_idempotent = {
-            use std::os::unix::fs::MetadataExt;
-            std::fs::metadata(&path).expect("segment exists").ino()
-        };
-        assert_eq!(
-            ino_after_create, ino_after_idempotent,
-            "idempotent retry must not unlink+recreate the live mirror"
-        );
-
-        // A stale mirror of a different size is still replaced (new
-        // inode, new size).
-        super::create_cross_pool_shmem(
-            &df,
-            "M",
-            &pool,
-            size + 1,
-            "torch.int64",
-            &[7_680_000],
-            "cpu",
-            None,
-        )
-        .expect("stale size replaced");
-        let meta = std::fs::metadata(&path).expect("segment exists");
-        let ino_after_replace = {
-            use std::os::unix::fs::MetadataExt;
-            meta.ino()
-        };
-        assert_ne!(
-            ino_after_create, ino_after_replace,
-            "stale-size mirror must be replaced with a fresh segment"
-        );
-        let _ = std::fs::remove_file(&path);
-    }
-}
-
-#[cfg(test)]
 mod debug_topic_tests {
     use super::rebuild_debug_topic_stream;
     use dora_message::metadata::{
@@ -13139,1063 +11377,5 @@ mod health_check_tests {
         let last = 10_000u64;
         let now = 1_000u64;
         assert!(!health_check_should_kill(true, last, now, TIMEOUT));
-    }
-}
-
-#[cfg(test)]
-mod cross_pool_write_tests {
-    use super::*;
-
-    /// Serializes tests that mutate the process-wide
-    /// `DORA_MEMORY_POOL_AUTH_TOKEN` env var. Rust tests run on parallel
-    /// threads; the auth-handshake tests key off that env on both the
-    /// send and verify sides, so a parallel `set_var` would flip the
-    /// handshake contract out from under them mid-flight.
-    static CROSS_DATA_AUTH_ENV_LOCK: std::sync::Mutex<()> = std::sync::Mutex::new(());
-
-    /// Panic-safe cleanup: unlink the test mirror from /dev/shm even when
-    /// an assertion fails mid-test (a leaked segment would pollute the
-    /// bench host's zero-residue checks).
-    struct ShmemCleanup(String);
-
-    impl Drop for ShmemCleanup {
-        fn drop(&mut self) {
-            let _ = std::fs::remove_file(format!("/dev/shm/{}", self.0));
-        }
-    }
-
-    /// Orphan sweep removes only segments under this machine's id prefix;
-    /// segments of other machines (sibling daemons on the same host) and
-    /// local (un-prefixed) pool segments survive.
-    #[test]
-    #[cfg(target_os = "linux")]
-    fn orphan_sweep_only_touches_own_machine_prefix() {
-        let dir = "/dev/shm";
-        // "orphanB" prefix isolates this test from the other cross-pool
-        // tests, which run in parallel and use segments under dora_pool_B_.
-        let own = "dora_pool_orphanB_orphantest_node_0";
-        let sibling = "dora_pool_orphanC_orphantest_node_0";
-        let local = "dora_pool_orphantest_node_0";
-        std::fs::write(format!("{dir}/{own}"), vec![0u8; 64]).unwrap();
-        std::fs::write(format!("{dir}/{sibling}"), vec![0u8; 64]).unwrap();
-        std::fs::write(format!("{dir}/{local}"), vec![0u8; 64]).unwrap();
-
-        let removed = cleanup_orphan_mirrors("orphanB");
-
-        assert_eq!(removed, 1);
-        assert!(!std::path::Path::new(&format!("{dir}/{own}")).exists());
-        assert!(std::path::Path::new(&format!("{dir}/{sibling}")).exists());
-        assert!(std::path::Path::new(&format!("{dir}/{local}")).exists());
-        // test hygiene
-        let _ = std::fs::remove_file(format!("{dir}/{sibling}"));
-        let _ = std::fs::remove_file(format!("{dir}/{local}"));
-    }
-
-    /// `finish_dataflow` must unlink the mirror segments this daemon
-    /// created for the finishing dataflow (dora-rs/dora#3194) — an
-    /// explicit `FreePool` never arrives when the dataflow ends by node
-    /// crash, `dora stop`, or abort — and must leave alone both the
-    /// origin-side entry (which owns no segment) and another dataflow's
-    /// still-live mirrors.
-    #[test]
-    #[cfg(target_os = "linux")]
-    fn finish_unlinks_own_mirrors_only() {
-        let finishing = Uuid::new_v4().to_string();
-        let other = Uuid::new_v4().to_string();
-
-        let name = |dataflow_id: &str, pool_id: &str| {
-            TensorPoolManager::cross_pool_shmem_name("finishB", dataflow_id, pool_id).unwrap()
-        };
-        // (dataflow, pool, is a mirror this daemon created, survives finish)
-        let fixture = [
-            (&finishing, "pool_node_0", true, false),
-            (&finishing, "pool_node_1", true, false),
-            // Origin-side entry of the finishing dataflow: no segment of
-            // its own, so the sender's live local pool must survive.
-            (&finishing, "pool_sender_0", false, true),
-            // Mirror of a dataflow that is still running.
-            (&other, "pool_node_0", true, true),
-        ];
-
-        let tensor_pool = TensorPoolManager::new();
-        let mut cleanup = Vec::new();
-        for (dataflow_id, pool_id, is_mirror, _) in fixture {
-            let shmem_name = name(dataflow_id, pool_id);
-            std::fs::write(format!("/dev/shm/{shmem_name}"), vec![0u8; 4096]).unwrap();
-            cleanup.push(ShmemCleanup(shmem_name.clone()));
-            tensor_pool.register_cross_pool(
-                dataflow_id.to_string(),
-                pool_id.to_string(),
-                "A".to_string(),
-                is_mirror.then_some(shmem_name),
-            );
-        }
-
-        assert_eq!(tensor_pool.cleanup_cross_pools(&finishing), 2);
-
-        for (dataflow_id, pool_id, _, survives) in fixture {
-            let path = format!("/dev/shm/{}", name(dataflow_id, pool_id));
-            assert_eq!(
-                std::path::Path::new(&path).exists(),
-                survives,
-                "{pool_id} of {dataflow_id}"
-            );
-        }
-    }
-
-    /// A stale mirror segment (leftover from a killed daemon that never
-    /// ran shutdown cleanup) must be replaced on re-register instead of
-    /// failing with EEXIST.
-    #[test]
-    #[cfg(target_os = "linux")]
-    fn stale_mirror_is_replaced_on_register() {
-        let dataflow_id = Uuid::new_v4();
-        let pool_id = "pool_node_0";
-        const SIZE: usize = 4096;
-        let shmem_name =
-            TensorPoolManager::cross_pool_shmem_name("B", &dataflow_id.to_string(), pool_id)
-                .unwrap();
-        // Simulate the leftover: a plain file under the mirror's name.
-        std::fs::write(format!("/dev/shm/{shmem_name}"), vec![0u8; 512]).unwrap();
-        let _cleanup = ShmemCleanup(shmem_name.clone());
-
-        create_cross_pool_shmem(
-            &dataflow_id,
-            "B",
-            pool_id,
-            SIZE,
-            "int64",
-            &[512],
-            "cpu",
-            None,
-        )
-        .unwrap();
-
-        // The recreated segment must be a valid DORADMA mirror.
-        let shmem = ShmemConf::new().os_id(&shmem_name).open().unwrap();
-        let magic = unsafe { std::slice::from_raw_parts(shmem.as_ptr(), 8) };
-        assert_eq!(magic, DORADMA_MAGIC);
-    }
-
-    #[test]
-    #[cfg(target_os = "linux")] // /dev/shm 段名（mirror 创建/校验）仅 Linux 有效
-    fn mirror_json_records_receiver_device() {
-        let dataflow_id = Uuid::new_v4();
-        let pool_id = "pool_node_0";
-        const SIZE: usize = 4096;
-        for device in ["cpu", "cuda:0"] {
-            let shmem_name =
-                TensorPoolManager::cross_pool_shmem_name("B", &dataflow_id.to_string(), pool_id)
-                    .unwrap();
-            let _cleanup = ShmemCleanup(shmem_name.clone());
-            create_cross_pool_shmem(
-                &dataflow_id,
-                "B",
-                pool_id,
-                SIZE,
-                "int64",
-                &[512],
-                device,
-                None,
-            )
-            .unwrap();
-            let shmem = ShmemConf::new().os_id(&shmem_name).open().unwrap();
-            let json_len = unsafe { read_header_u64(shmem.as_ptr().add(8)) } as usize;
-            let json_bytes = unsafe {
-                std::slice::from_raw_parts(shmem.as_ptr().add(DORADMA_HEADER_SIZE), json_len)
-            };
-            let json = String::from_utf8(json_bytes.to_vec()).unwrap();
-            assert!(
-                json.contains(&format!("\"pinned_type\":\"{device}\"")),
-                "mirror json {json} must record receiver device {device}"
-            );
-            std::fs::remove_file(format!("/dev/shm/{shmem_name}")).unwrap();
-        }
-    }
-
-    /// Concurrent writers to the same mirror must not interleave bytes:
-    /// after every round the data region holds one writer's complete
-    /// pattern, never a mixture. Without the per-pool lock, overlapping
-    /// memcpys of distinct fill bytes interleave and the final frame
-    /// passes the seqlock (even generation) with mixed bytes.
-    #[test]
-    #[cfg(target_os = "linux")]
-    fn concurrent_mirror_writes_never_interleave_bytes() {
-        let dataflow_id = Uuid::new_v4();
-        let pool_id = "pool_node_0";
-        const SIZE: usize = 4 * 1024 * 1024;
-        create_cross_pool_shmem(
-            &dataflow_id,
-            "B",
-            pool_id,
-            SIZE,
-            "int64",
-            &[8192],
-            "cpu",
-            None,
-        )
-        .unwrap();
-        let shmem_name =
-            TensorPoolManager::cross_pool_shmem_name("B", &dataflow_id.to_string(), pool_id)
-                .unwrap();
-        let _cleanup = ShmemCleanup(shmem_name.clone());
-        let shmem = ShmemConf::new().os_id(&shmem_name).open().unwrap();
-        // data_offset comes from the header (json_len varies), not a constant.
-        let data_offset = unsafe { read_header_u64(shmem.as_ptr().add(16)) as usize };
-
-        // Writers with distinct fill bytes race the same mirror.
-        // Patterns are allocated once and reused across rounds.
-        const WRITERS: u8 = 8;
-        let patterns: Vec<Vec<u8>> = (0..WRITERS).map(|w| vec![w; SIZE]).collect();
-        for round in 0..200 {
-            std::thread::scope(|scope| {
-                for pattern in patterns.iter() {
-                    let pattern = pattern.as_slice();
-                    scope.spawn(move || {
-                        write_cross_pool_data(&dataflow_id, "B", pool_id, pattern, SIZE);
-                    });
-                }
-            });
-            // The data region must hold exactly one writer's full pattern.
-            let first = unsafe { *shmem.as_ptr().add(data_offset) };
-            let data = unsafe { std::slice::from_raw_parts(shmem.as_ptr().add(data_offset), SIZE) };
-            assert!(
-                data.iter().all(|b| *b == first),
-                "round {round}: interleaved bytes in mirror data"
-            );
-            // Seqlock: generation is even (complete) after the round.
-            let generation =
-                unsafe { std::ptr::read_volatile(shmem.as_ptr().add(96) as *const u64) };
-            assert_eq!(
-                generation % 2,
-                0,
-                "round {round}: odd generation after write"
-            );
-        }
-    }
-
-    /// The write-commit ack resolves only the seq-matched pending reply.
-    ///
-    /// The same-host cross-daemon smoke test reads via the `direct ==
-    /// true` path and bypasses the MemoryPoolWrite/MemoryPoolWriteAck
-    /// machinery entirely (only the manual two-host runs exercise it), so
-    /// this test pins the ack semantics directly: a stale ack (a previous
-    /// write's seq) must not resolve anything, the seq-matched ack
-    /// resolves exactly its own pending reply, and a failed mirror write
-    /// surfaces as an error reply.
-    #[test]
-    fn write_ack_resolves_only_seq_matched_pending_reply() {
-        use tokio::sync::oneshot;
-
-        let df = Uuid::new_v4();
-        let pool = "pool_sender_node_1".to_string();
-
-        // Two in-flight writes to the same pool: seq 1 then seq 2.
-        let (tx1, mut rx1) = oneshot::channel();
-        let (tx2, mut rx2) = oneshot::channel();
-        {
-            let mut pending = CROSS_WRITE_PENDING
-                .lock()
-                .unwrap_or_else(|e| e.into_inner());
-            pending.insert((df, pool.clone(), 1), tx1);
-            pending.insert((df, pool.clone(), 2), tx2);
-        }
-
-        // A stale ack (a previous write's seq) resolves nothing.
-        assert!(!resolve_cross_write_ack(df, pool.clone(), 0, true, None));
-        assert!(matches!(
-            rx1.try_recv(),
-            Err(oneshot::error::TryRecvError::Empty)
-        ));
-        assert!(matches!(
-            rx2.try_recv(),
-            Err(oneshot::error::TryRecvError::Empty)
-        ));
-
-        // The seq-matched ack resolves exactly its own pending reply.
-        assert!(resolve_cross_write_ack(df, pool.clone(), 1, true, None));
-        match rx1.try_recv() {
-            Ok(DaemonReply::Result(Ok(()))) => {}
-            other => panic!("seq-1 ack resolved the wrong reply: {other:?}"),
-        }
-        // seq 2 is untouched by the seq-1 ack.
-        assert!(matches!(
-            rx2.try_recv(),
-            Err(oneshot::error::TryRecvError::Empty)
-        ));
-
-        // A failed mirror write surfaces as an error reply.
-        assert!(resolve_cross_write_ack(
-            df,
-            pool.clone(),
-            2,
-            false,
-            Some("mirror segment missing".to_string()),
-        ));
-        match rx2.try_recv() {
-            Ok(DaemonReply::Result(Err(e))) => {
-                assert!(e.contains("mirror segment missing"));
-            }
-            other => panic!("failed write must resolve to an error reply: {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn write_lock_is_dataflow_scoped() {
-        let df1 = Uuid::new_v4();
-        let df2 = Uuid::new_v4();
-        let pool = "pool_sender_node_1".to_string();
-        let mut locks = CROSS_POOL_WRITE_LOCKS_ASYNC.lock().await;
-        let mut entry = |df, pool: &str| {
-            locks
-                .entry((df, pool.to_string()))
-                .or_insert_with(|| std::sync::Arc::new(tokio::sync::Mutex::new(())))
-                .clone()
-        };
-        // Same pool id in two dataflows → distinct locks (concurrent
-        // dataflows never serialize on each other's segments).
-        let l1 = entry(df1, &pool);
-        let l2 = entry(df2, &pool);
-        assert!(!std::sync::Arc::ptr_eq(&l1, &l2));
-        // Same dataflow, same pool → the same lock (writes to one segment
-        // stay serialized).
-        assert!(std::sync::Arc::ptr_eq(&l1, &entry(df1, &pool)));
-        // Same dataflow, different pool → distinct locks.
-        assert!(!std::sync::Arc::ptr_eq(
-            &l1,
-            &entry(df1, "pool_sender_node_2")
-        ));
-    }
-
-    /// The direct-TCP fallback warns exactly once per pool: repeated
-    /// failures while degraded stay silent, and recovery is reported once.
-    ///
-    /// A broken link is the steady state for the zenoh fallback — a
-    /// per-frame warn would flood the daemon log for the whole outage.
-    #[test]
-    fn direct_fallback_warns_once_per_pool_and_reports_recovery() {
-        let df = Uuid::new_v4();
-        let pool = "pool_sender_node_1".to_string();
-
-        // First failure: newly degraded, the caller warns.
-        assert!(note_direct_degraded(df, &pool));
-        // Further failures while degraded: silent.
-        assert!(!note_direct_degraded(df, &pool));
-        assert!(!note_direct_degraded(df, &pool));
-        // Recovery: was degraded, the caller logs it once.
-        assert!(note_direct_recovered(df, &pool));
-        // Recovery without being degraded: nothing to report.
-        assert!(!note_direct_recovered(df, &pool));
-
-        // Pools are tracked independently.
-        assert!(note_direct_degraded(df, "pool_sender_node_2"));
-        assert!(!note_direct_degraded(df, "pool_sender_node_2"));
-        assert!(note_direct_recovered(df, "pool_sender_node_2"));
-        // A fresh dataflow never collides (keyed by UUID).
-        assert!(note_direct_degraded(Uuid::new_v4(), &pool));
-    }
-
-    /// Build a direct-TCP frame header (magic + dataflow + pool + seq +
-    /// size) with no payload bytes.
-    fn build_frame_header(dataflow_id: Uuid, pool_id: &str, seq: u64, size: u64) -> Vec<u8> {
-        let mut frame = Vec::new();
-        frame.extend_from_slice(&CROSS_DATA_MAGIC.to_be_bytes());
-        frame.extend_from_slice(dataflow_id.as_bytes());
-        let pool_bytes = pool_id.as_bytes();
-        frame.extend_from_slice(&(pool_bytes.len() as u32).to_be_bytes());
-        frame.extend_from_slice(pool_bytes);
-        frame.extend_from_slice(&seq.to_be_bytes());
-        frame.extend_from_slice(&size.to_be_bytes());
-        frame
-    }
-
-    /// A wire-controlled `size` near `u64::MAX` must be rejected, not
-    /// wrap `data_offset + size` past the bounds check (which would
-    /// construct a `size`-byte slice past the mapping — UB, and a
-    /// remote-triggerable abort in debug builds).
-    #[tokio::test]
-    #[cfg(target_os = "linux")]
-    async fn wire_size_overflow_is_rejected_not_aborted() {
-        use tokio::io::AsyncWriteExt;
-
-        let dataflow_id = Uuid::new_v4();
-        let pool_id = "pool_node_0";
-        const SIZE: usize = 64 * 1024;
-        create_cross_pool_shmem(
-            &dataflow_id,
-            "B",
-            pool_id,
-            SIZE,
-            "int64",
-            &[8192],
-            "cpu",
-            None,
-        )
-        .unwrap();
-        let shmem_name =
-            TensorPoolManager::cross_pool_shmem_name("B", &dataflow_id.to_string(), pool_id)
-                .unwrap();
-        let _cleanup = ShmemCleanup(shmem_name.clone());
-        let tensor_pool = TensorPoolManager::new();
-        tensor_pool.register_cross_pool(
-            dataflow_id.to_string(),
-            pool_id.to_string(),
-            "A".to_string(),
-            Some(shmem_name.clone()),
-        );
-
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-        let server = tokio::spawn(async move {
-            let (mut stream, _) = listener.accept().await.unwrap();
-            handle_cross_data_frame(&mut stream, &tensor_pool, "B").await
-        });
-
-        // data_offset + u64::MAX wraps to a small value without the
-        // checked add — the frame must be rejected with the ack info.
-        let frame = build_frame_header(dataflow_id, pool_id, 1, u64::MAX);
-        let mut client = tokio::net::TcpStream::connect(addr).await.unwrap();
-        client.write_all(&frame).await.unwrap();
-
-        let err = server.await.unwrap().unwrap_err();
-        assert!(
-            err.ack == Some((dataflow_id, pool_id.to_string(), 1)),
-            "overflow rejection must carry the ack info: {err:?}"
-        );
-        // Rejection can come from the registered-size check (newer,
-        // fires first: u64::MAX != registered size) or the checked-add
-        // overflow guard — either way the wire-controlled size is
-        // refused without aborting.
-        assert!(
-            err.message.contains("overflow") || err.message.contains("registered pool size"),
-            "expected size rejection, got: {}",
-            err.message
-        );
-        // The mirror's seqlock generation is untouched: no writer began,
-        // so it stays at the initial odd (in-progress) value 1.
-        let shmem = ShmemConf::new().os_id(&shmem_name).open().unwrap();
-        let generation = unsafe { std::ptr::read_volatile(shmem.as_ptr().add(96) as *const u64) };
-        assert_eq!(
-            generation, 1,
-            "generation must be untouched (no write began)"
-        );
-    }
-
-    /// The direct-TCP serve layer fails the origin fast on a bad frame:
-    /// `serve_cross_data_frame` publishes `MemoryPoolWriteAck { ok: false }`
-    /// over zenoh when the frame's identity is parsed but the write cannot
-    /// proceed. This closes the last untested half of the standing
-    /// non-blocker — the codec round-trip covers the happy path, this
-    /// covers the error path's zenoh ack publish (only the true two-host
-    /// transfer remains manual).
-    #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-    #[cfg(target_os = "linux")]
-    async fn serve_error_publishes_failed_write_ack() {
-        use tokio::io::AsyncWriteExt;
-
-        // Hermetic zenoh pair: mirror listens, origin dials; no scouting.
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let port = listener.local_addr().unwrap().port();
-        drop(listener);
-        let mut mirror_cfg = zenoh::Config::default();
-        let mut origin_cfg = zenoh::Config::default();
-        for cfg in [&mut mirror_cfg, &mut origin_cfg] {
-            cfg.insert_json5("scouting/multicast/enabled", "false")
-                .unwrap();
-            cfg.insert_json5("scouting/gossip/enabled", "false")
-                .unwrap();
-        }
-        mirror_cfg
-            .insert_json5(
-                "listen/endpoints",
-                &format!(r#"{{ peer: ["tcp/127.0.0.1:{port}"] }}"#),
-            )
-            .unwrap();
-        mirror_cfg
-            .insert_json5("listen/exit_on_failure", "false")
-            .unwrap();
-        origin_cfg
-            .insert_json5(
-                "connect/endpoints",
-                &format!(r#"{{ peer: ["tcp/127.0.0.1:{port}"] }}"#),
-            )
-            .unwrap();
-        let mirror_session = zenoh::open(mirror_cfg).await.unwrap();
-        let origin_session = zenoh::open(origin_cfg).await.unwrap();
-
-        // Mirror pool the frame will target.
-        let dataflow_id = Uuid::new_v4();
-        let pool_id = "pool_node_0";
-        const SIZE: usize = 64 * 1024;
-        create_cross_pool_shmem(
-            &dataflow_id,
-            "B",
-            pool_id,
-            SIZE,
-            "int64",
-            &[8192],
-            "cpu",
-            None,
-        )
-        .unwrap();
-        let shmem_name =
-            TensorPoolManager::cross_pool_shmem_name("B", &dataflow_id.to_string(), pool_id)
-                .unwrap();
-        let _cleanup = ShmemCleanup(shmem_name.clone());
-        let tensor_pool = TensorPoolManager::new();
-        tensor_pool.register_cross_pool(
-            dataflow_id.to_string(),
-            pool_id.to_string(),
-            "A".to_string(),
-            Some(shmem_name.clone()),
-        );
-
-        // Data listener served by the mirror session's process context.
-        let data_listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let data_addr = data_listener.local_addr().unwrap();
-        let clock = Arc::new(HLC::default());
-        let serve = tokio::spawn(async move {
-            let (mut stream, _) = data_listener.accept().await.unwrap();
-            serve_cross_data_frame(
-                &tensor_pool,
-                "B",
-                &mirror_session,
-                &clock,
-                None,
-                &mut stream,
-            )
-            .await
-        });
-
-        // Origin-side subscriber on the memory-pool topic (the daemon's
-        // per-dataflow loop in miniature).
-        let topic = dataflow_memory_pool_topic(&dataflow_id);
-        let subscriber = origin_session.declare_subscriber(&topic).await.unwrap();
-        // The serve layer publishes the failed-write ack exactly once
-        // (event-driven, no retry — fine in production, where the
-        // subscription is established long before any write). Give the
-        // subscription interest time to propagate to the mirror session,
-        // or the single ack put would be dropped into the void.
-        tokio::time::sleep(std::time::Duration::from_millis(300)).await;
-
-        // A bad frame (size near u64::MAX → checked-add rejection) with a
-        // parsed identity: the serve layer must publish ok=false.
-        let mut client = tokio::net::TcpStream::connect(data_addr).await.unwrap();
-        let frame = build_frame_header(dataflow_id, pool_id, 1, u64::MAX);
-        client.write_all(&frame).await.unwrap();
-
-        let serve_result = serve.await.unwrap();
-        assert!(serve_result.is_err(), "serve must report the frame error");
-
-        // The origin receives the failed-write ack with matching identity.
-        let mut ack_received = false;
-        for _ in 0..10 {
-            match tokio::time::timeout(std::time::Duration::from_secs(2), subscriber.recv_async())
-                .await
-            {
-                Ok(Ok(sample)) => {
-                    let bytes = sample.payload().to_bytes();
-                    let event =
-                        Timestamped::<InterDaemonEvent>::deserialize_inter_daemon_event(&bytes)
-                            .unwrap();
-                    match event.inner {
-                        InterDaemonEvent::MemoryPoolWriteAck {
-                            dataflow_id: df,
-                            shared_memory_id,
-                            seq,
-                            ok,
-                            error,
-                        } => {
-                            assert_eq!(df, dataflow_id);
-                            assert_eq!(shared_memory_id, pool_id);
-                            assert_eq!(seq, 1);
-                            assert!(!ok, "failed write must ack ok=false");
-                            assert!(
-                                error.is_some()
-                                    && (error.as_deref().unwrap().contains("overflow")
-                                        || error
-                                            .as_deref()
-                                            .unwrap()
-                                            .contains("registered pool size")),
-                                "failed write ack must carry the error: {error:?}"
-                            );
-                            ack_received = true;
-                        }
-                        other => panic!("unexpected event: {other:?}"),
-                    }
-                }
-                Ok(Err(e)) => panic!("subscriber closed: {e}"),
-                Err(_) => {} // interest not yet propagated; keep waiting
-            }
-            if ack_received {
-                break;
-            }
-        }
-        assert!(ack_received, "failed-write ack never arrived over zenoh");
-    }
-
-    /// A payload read that fails mid-frame (TCP drop) leaves the seqlock
-    /// generation odd — readers reject the torn frame and never see
-    /// half-written bytes; the next full write self-heals — and the
-    /// error carries the ack info so the origin fails fast instead of
-    /// waiting out the commit-ack timeout.
-    #[tokio::test]
-    #[cfg(target_os = "linux")]
-    async fn payload_read_failure_stays_odd_and_carries_ack() {
-        use tokio::io::AsyncWriteExt;
-
-        let dataflow_id = Uuid::new_v4();
-        let pool_id = "pool_node_0";
-        const SIZE: usize = 64 * 1024;
-        create_cross_pool_shmem(
-            &dataflow_id,
-            "B",
-            pool_id,
-            SIZE,
-            "int64",
-            &[8192],
-            "cpu",
-            None,
-        )
-        .unwrap();
-        let shmem_name =
-            TensorPoolManager::cross_pool_shmem_name("B", &dataflow_id.to_string(), pool_id)
-                .unwrap();
-        let _cleanup = ShmemCleanup(shmem_name.clone());
-        let tensor_pool = TensorPoolManager::new();
-        tensor_pool.register_cross_pool(
-            dataflow_id.to_string(),
-            pool_id.to_string(),
-            "A".to_string(),
-            Some(shmem_name.clone()),
-        );
-
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-        let tensor_pool_1 = tensor_pool.clone();
-        let server = tokio::spawn(async move {
-            let (mut stream, _) = listener.accept().await.unwrap();
-            handle_cross_data_frame(&mut stream, &tensor_pool_1, "B").await
-        });
-        let mut client = tokio::net::TcpStream::connect(addr).await.unwrap();
-
-        // Frame 1: a full, valid write — the generation advances to an
-        // even baseline (2) that the failure must leave odd-on-top-of.
-        let mut frame1 = build_frame_header(dataflow_id, pool_id, 1, SIZE as u64);
-        frame1.extend(std::iter::repeat_n(7u8, SIZE));
-        client.write_all(&frame1).await.unwrap();
-        let ack1 = server.await.unwrap().unwrap().unwrap();
-        assert_eq!(ack1.2, 1);
-        let shmem = ShmemConf::new().os_id(&shmem_name).open().unwrap();
-        let generation = unsafe { std::ptr::read_volatile(shmem.as_ptr().add(96) as *const u64) };
-        assert_eq!(
-            generation, 2,
-            "frame 1 must complete the write (even gen 2)"
-        );
-
-        // Frame 2: header + half the payload, then the connection drops —
-        // the read fails mid-frame.
-        let listener2 = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr2 = listener2.local_addr().unwrap();
-        let tensor_pool_2 = tensor_pool.clone();
-        let server2 = tokio::spawn(async move {
-            let (mut stream, _) = listener2.accept().await.unwrap();
-            handle_cross_data_frame(&mut stream, &tensor_pool_2, "B").await
-        });
-        let mut client2 = tokio::net::TcpStream::connect(addr2).await.unwrap();
-        let mut frame2 = build_frame_header(dataflow_id, pool_id, 2, SIZE as u64);
-        frame2.extend(std::iter::repeat_n(9u8, SIZE / 2));
-        client2.write_all(&frame2).await.unwrap();
-        // Drop the connection: the mirror's read_exact fails with EOF.
-        drop(client2);
-
-        let err = server2.await.unwrap().unwrap_err();
-        assert!(
-            err.ack == Some((dataflow_id, pool_id.to_string(), 2)),
-            "mid-frame read failure must carry the ack info: {err:?}"
-        );
-        assert!(err.message.contains("read payload"), "got: {}", err.message);
-
-        // The generation stays odd (in-progress) — the fail-safe: readers
-        // reject the torn frame rather than reading half-written bytes.
-        let generation = unsafe { std::ptr::read_volatile(shmem.as_ptr().add(96) as *const u64) };
-        assert_eq!(generation, 3, "generation must stay odd on the torn frame");
-        // A subsequent full write self-heals: begin keeps the odd
-        // generation, writes the frame, and publishes the even one.
-        let listener3 = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr3 = listener3.local_addr().unwrap();
-        let tensor_pool_3 = tensor_pool.clone();
-        let server3 = tokio::spawn(async move {
-            let (mut stream, _) = listener3.accept().await.unwrap();
-            handle_cross_data_frame(&mut stream, &tensor_pool_3, "B").await
-        });
-        let mut client3 = tokio::net::TcpStream::connect(addr3).await.unwrap();
-        let mut frame3 = build_frame_header(dataflow_id, pool_id, 3, SIZE as u64);
-        frame3.extend(std::iter::repeat_n(11u8, SIZE));
-        client3.write_all(&frame3).await.unwrap();
-        let ack3 = server3.await.unwrap().unwrap().unwrap();
-        assert_eq!(ack3.2, 3);
-        let generation = unsafe { std::ptr::read_volatile(shmem.as_ptr().add(96) as *const u64) };
-        assert_eq!(
-            generation, 4,
-            "the next full write must self-heal to an even generation"
-        );
-        let data_offset = unsafe { read_header_u64(shmem.as_ptr().add(16)) } as usize;
-        let data = unsafe { std::slice::from_raw_parts(shmem.as_ptr().add(data_offset), SIZE) };
-        assert!(
-            data.iter().all(|b| *b == 11),
-            "self-healed frame must be fully visible"
-        );
-    }
-
-    /// The zenoh ack publish itself: the mirror-side publish helper over a
-    /// real zenoh transport (loopback TCP), received and deserialized by
-    /// the origin-side subscriber, resolving the seq-matched pending reply.
-    ///
-    /// `write_ack_resolves_only_seq_matched_pending_reply` pins the
-    /// resolver in isolation; this test covers the publish → transport →
-    /// subscribe → deserialize chain that feeds it — the last uncovered
-    /// link of the commit protocol (only the true two-host transfer still
-    /// needs manual runs). Two sessions are required because the mirror
-    /// publishes with `Locality::Remote`: a same-session subscriber would
-    /// never receive its own put.
-    #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-    async fn zenoh_ack_publish_resolves_pending_reply() {
-        // Free loopback port for the mirror session's listener.
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let port = listener.local_addr().unwrap().port();
-        drop(listener);
-
-        let mut mirror_cfg = zenoh::Config::default();
-        let mut origin_cfg = zenoh::Config::default();
-        for cfg in [&mut mirror_cfg, &mut origin_cfg] {
-            // Hermetic: no scouting, so the test can neither touch nor be
-            // touched by other zenoh instances on the host.
-            cfg.insert_json5("scouting/multicast/enabled", "false")
-                .unwrap();
-            cfg.insert_json5("scouting/gossip/enabled", "false")
-                .unwrap();
-        }
-        mirror_cfg
-            .insert_json5(
-                "listen/endpoints",
-                &format!(r#"{{ peer: ["tcp/127.0.0.1:{port}"] }}"#),
-            )
-            .unwrap();
-        mirror_cfg
-            .insert_json5("listen/exit_on_failure", "false")
-            .unwrap();
-        origin_cfg
-            .insert_json5(
-                "connect/endpoints",
-                &format!(r#"{{ peer: ["tcp/127.0.0.1:{port}"] }}"#),
-            )
-            .unwrap();
-        let mirror_session = zenoh::open(mirror_cfg).await.unwrap();
-        let origin_session = zenoh::open(origin_cfg).await.unwrap();
-
-        let df = Uuid::new_v4();
-        let pool = "pool_sender_node_1".to_string();
-        let seq = 7u64;
-        let topic = dataflow_memory_pool_topic(&df);
-
-        // Origin-side subscriber, mirroring the daemon's per-dataflow loop.
-        let subscriber = origin_session.declare_subscriber(&topic).await.unwrap();
-
-        // The write this ack commits is pending its reply.
-        let (tx, rx) = tokio::sync::oneshot::channel();
-        CROSS_WRITE_PENDING
-            .lock()
-            .unwrap_or_else(|e| e.into_inner())
-            .insert((df, pool.clone(), seq), tx);
-
-        // Mirror-side publish through the production helper. Retry the put
-        // until the ack arrives: the subscriber interest must propagate to
-        // the mirror session over the fresh link, and a put that precedes
-        // the interest is dropped (Block congestion control never drops, so
-        // retries converge).
-        let clock = Arc::new(HLC::default());
-        let mut ack_received = false;
-        for _ in 0..10 {
-            publish_memory_pool_event(
-                &mirror_session,
-                &clock,
-                &df,
-                &InterDaemonEvent::MemoryPoolWriteAck {
-                    dataflow_id: df,
-                    shared_memory_id: pool.clone(),
-                    seq,
-                    ok: true,
-                    error: None,
-                },
-                None,
-            )
-            .await
-            .unwrap();
-            match tokio::time::timeout(std::time::Duration::from_secs(2), subscriber.recv_async())
-                .await
-            {
-                Ok(Ok(sample)) => {
-                    // Deserialize with the production method and resolve.
-                    let bytes = sample.payload().to_bytes();
-                    let event =
-                        Timestamped::<InterDaemonEvent>::deserialize_inter_daemon_event(&bytes)
-                            .unwrap();
-                    match event.inner {
-                        InterDaemonEvent::MemoryPoolWriteAck {
-                            dataflow_id,
-                            shared_memory_id,
-                            seq: ack_seq,
-                            ok,
-                            error,
-                        } => {
-                            assert_eq!(dataflow_id, df);
-                            assert_eq!(shared_memory_id, pool);
-                            assert_eq!(ack_seq, seq);
-                            assert!(ok);
-                            assert!(error.is_none());
-                            assert!(resolve_cross_write_ack(
-                                dataflow_id,
-                                shared_memory_id,
-                                ack_seq,
-                                ok,
-                                error,
-                            ));
-                        }
-                        other => panic!("unexpected event on memory-pool topic: {other:?}"),
-                    }
-                    ack_received = true;
-                }
-                Ok(Err(e)) => panic!("memory-pool subscriber closed: {e}"),
-                Err(_) => {} // interest not yet propagated; put again
-            }
-            if ack_received {
-                break;
-            }
-        }
-        assert!(ack_received, "ack never arrived over the zenoh link");
-
-        // The pending write's reply is resolved by the zenoh-delivered ack.
-        match rx.await {
-            Ok(DaemonReply::Result(Ok(()))) => {}
-            other => panic!("pending reply not resolved by the zenoh ack: {other:?}"),
-        }
-    }
-
-    /// The direct-TCP data-plane codec: a frame sent via
-    /// `send_cross_data_frame` over a loopback connection is parsed by
-    /// `handle_cross_data_frame` and written straight into the mirror
-    /// segment — payload bytes land in the data region under the seqlock,
-    /// and the returned ack info matches (dataflow, pool, seq). This is
-    /// the new steady-state cross-machine write path, which the same-host
-    /// smoke (direct == true) bypasses entirely.
-    #[test]
-    #[cfg(target_os = "linux")]
-    fn direct_tcp_frame_round_trip_writes_mirror() {
-        // This test exercises the token-less contract (both ends skip the
-        // handshake when no token is configured). Serialize against the
-        // auth test's env mutation and clear the variable defensively so
-        // a token in the developer's shell cannot flip this test into
-        // handshake mode mid-flight.
-        //
-        // SAFETY (env mutation in tests): `CROSS_DATA_AUTH_ENV_LOCK`
-        // serializes every test that reads or writes this variable; the
-        // guard runs on this test's own thread and is never awaited on.
-        let _env_lock = CROSS_DATA_AUTH_ENV_LOCK.lock().unwrap();
-        unsafe { std::env::remove_var("DORA_MEMORY_POOL_AUTH_TOKEN") };
-        struct TokenEnvGuard;
-        impl Drop for TokenEnvGuard {
-            fn drop(&mut self) {
-                unsafe { std::env::remove_var("DORA_MEMORY_POOL_AUTH_TOKEN") };
-            }
-        }
-        let _token_guard = TokenEnvGuard;
-
-        let runtime = tokio::runtime::Builder::new_current_thread()
-            .enable_all()
-            .build()
-            .unwrap();
-        runtime.block_on(async {
-            let dataflow_id = Uuid::new_v4();
-            let pool_id = "pool_node_0";
-            const SIZE: usize = 64 * 1024;
-            create_cross_pool_shmem(
-                &dataflow_id,
-                "B",
-                pool_id,
-                SIZE,
-                "int64",
-                &[8192],
-                "cpu",
-                None,
-            )
-            .unwrap();
-            let shmem_name =
-                TensorPoolManager::cross_pool_shmem_name("B", &dataflow_id.to_string(), pool_id)
-                    .unwrap();
-            let _cleanup = ShmemCleanup(shmem_name.clone());
-            let tensor_pool = TensorPoolManager::new();
-            tensor_pool.register_cross_pool(
-                dataflow_id.to_string(),
-                pool_id.to_string(),
-                "A".to_string(),
-                Some(shmem_name.clone()),
-            );
-
-            let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-            let addr = listener.local_addr().unwrap();
-            let server = tokio::spawn(async move {
-                let (mut stream, _) = listener.accept().await.unwrap();
-                handle_cross_data_frame(&mut stream, &tensor_pool, "B").await
-            });
-            let payload: Vec<u8> = (0..SIZE).map(|i| (i % 251) as u8).collect();
-            let conns = Arc::new(std::sync::Mutex::new(HashMap::new()));
-            send_cross_data_frame(&conns, addr, dataflow_id, pool_id, 42, &payload)
-                .await
-                .unwrap();
-            let ack_info = server.await.unwrap().unwrap().unwrap();
-            assert_eq!(ack_info.0, dataflow_id);
-            assert_eq!(ack_info.1, pool_id);
-            assert_eq!(ack_info.2, 42);
-
-            let shmem = ShmemConf::new().os_id(&shmem_name).open().unwrap();
-            let data_offset = unsafe { read_header_u64(shmem.as_ptr().add(16)) } as usize;
-            let data = unsafe { std::slice::from_raw_parts(shmem.as_ptr().add(data_offset), SIZE) };
-            assert_eq!(
-                data,
-                payload.as_slice(),
-                "mirror data region must equal the payload"
-            );
-            // Seqlock: generation is even (complete) after the write.
-            let generation =
-                unsafe { std::ptr::read_volatile(shmem.as_ptr().add(96) as *const u64) };
-            assert_eq!(generation % 2, 0, "odd generation after write");
-        });
-    }
-
-    /// The direct-TCP auth handshake: with `DORA_MEMORY_POOL_AUTH_TOKEN`
-    /// set, a sender with the shared token is accepted and its frame
-    /// served; a sender without it (or with the wrong token) is rejected
-    /// before any frame is parsed.
-    #[tokio::test]
-    async fn auth_handshake_accepts_shared_token_and_rejects_others() {
-        // Hermetic env for this test: set and restore.
-        //
-        // SAFETY (env mutation in tests): `CROSS_DATA_AUTH_ENV_LOCK`
-        // serializes every test that reads or writes this variable, so no
-        // other test thread can observe the mutation mid-test. The std
-        // mutex is only ever locked briefly by this test's own thread
-        // (never awaited on), so it cannot deadlock the async body.
-        let _env_lock = CROSS_DATA_AUTH_ENV_LOCK.lock().unwrap();
-        unsafe { std::env::set_var("DORA_MEMORY_POOL_AUTH_TOKEN", "test-shared-secret") };
-        struct EnvGuard;
-        impl Drop for EnvGuard {
-            fn drop(&mut self) {
-                unsafe { std::env::remove_var("DORA_MEMORY_POOL_AUTH_TOKEN") };
-            }
-        }
-        let _guard = EnvGuard;
-
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-        let server = tokio::spawn(async move {
-            let (mut stream, _) = listener.accept().await.unwrap();
-            match tokio::time::timeout(
-                std::time::Duration::from_secs(5),
-                auth_handshake_verify(&mut stream),
-            )
-            .await
-            {
-                Ok(result) => result,
-                Err(_) => Err("auth handshake timed out".to_string()),
-            }
-        });
-
-        // Correct token: accepted.
-        let mut client = tokio::net::TcpStream::connect(addr).await.unwrap();
-        auth_handshake_send(&mut client, "test-shared-secret")
-            .await
-            .unwrap();
-        server.await.unwrap().unwrap();
-        assert_eq!((), (), "shared token must be accepted");
-
-        // Wrong token: rejected on the verify side and an error on the
-        // send side (the peer answers AUTH_FAIL).
-        let listener2 = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr2 = listener2.local_addr().unwrap();
-        let server2 = tokio::spawn(async move {
-            let (mut stream, _) = listener2.accept().await.unwrap();
-            tokio::time::timeout(
-                std::time::Duration::from_secs(5),
-                auth_handshake_verify(&mut stream),
-            )
-            .await
-        });
-        let mut client2 = tokio::net::TcpStream::connect(addr2).await.unwrap();
-        let send_err = auth_handshake_send(&mut client2, "wrong-token")
-            .await
-            .expect_err("wrong token must fail the handshake");
-        assert!(
-            send_err.contains("rejected"),
-            "send side must report the rejection: {send_err}"
-        );
-        let verify_err = server2.await.unwrap().unwrap().unwrap_err();
-        assert!(
-            verify_err.contains("mismatch"),
-            "verify side must report the mismatch: {verify_err}"
-        );
-    }
-
-    /// The listener's auth gate, driven through a real socket: a peer
-    /// with the wrong token must be rejected (no frame may be served)
-    /// and a peer with the shared token must pass. This pins the
-    /// nested-result flatten in `cross_data_auth_gate` — a regression
-    /// there (binding only the outer timeout, as the listener did
-    /// before) would let a mismatched peer through, while the
-    /// direct-call tests of `auth_handshake_verify` would stay green.
-    #[tokio::test]
-    async fn auth_gate_rejects_wrong_token_through_real_socket() {
-        let _env_lock = CROSS_DATA_AUTH_ENV_LOCK.lock().unwrap();
-        unsafe { std::env::set_var("DORA_MEMORY_POOL_AUTH_TOKEN", "test-shared-secret") };
-        struct EnvGuard;
-        impl Drop for EnvGuard {
-            fn drop(&mut self) {
-                unsafe { std::env::remove_var("DORA_MEMORY_POOL_AUTH_TOKEN") };
-            }
-        }
-        let _guard = EnvGuard;
-
-        // Wrong token: the gate must reject, mirroring the listener's
-        // drop-before-serving behavior.
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-        let server = tokio::spawn(async move {
-            let (mut stream, _) = listener.accept().await.unwrap();
-            cross_data_auth_gate(&mut stream).await
-        });
-        let mut client = tokio::net::TcpStream::connect(addr).await.unwrap();
-        auth_handshake_send(&mut client, "wrong-token")
-            .await
-            .expect_err("wrong token must fail the send side");
-        let gate_err = server.await.unwrap().unwrap_err();
-        assert!(
-            gate_err.contains("mismatch"),
-            "gate must reject the mismatched peer: {gate_err}"
-        );
-
-        // Correct token: the gate must pass, exactly as the listener
-        // would then serve frames.
-        let listener2 = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr2 = listener2.local_addr().unwrap();
-        let server2 = tokio::spawn(async move {
-            let (mut stream, _) = listener2.accept().await.unwrap();
-            cross_data_auth_gate(&mut stream).await
-        });
-        let mut client2 = tokio::net::TcpStream::connect(addr2).await.unwrap();
-        auth_handshake_send(&mut client2, "test-shared-secret")
-            .await
-            .unwrap();
-        server2.await.unwrap().unwrap();
-        assert_eq!((), (), "shared token must pass the gate");
     }
 }
